refactor(matrix): address submodule review polish
Verification records now preserve raw Matrix user IDs alongside sanitized wire/display IDs so SDK lookups and same-peer eviction use protocol identity, while raw fields remain serde-skipped. Matrix submodule tests are co-located with inbound DLQ, recovery, and verification code, DLQ wire constants live with the DLQ codec, and module headers document each moved seam.

The wire guard now requires src/channels/matrix.rs explicitly, concatenates production Matrix submodules for alias/macro guard coverage, and keeps self-test fixtures aligned with the split. Validation: cargo-serial fmt --check; cargo-serial check --all-targets; cargo-serial clippy --all-targets --all-features -- -D warnings; cargo-serial nextest run matrix_inbound_dlq; cargo-serial nextest run recovery_key; cargo-serial nextest run verification; bash -n + normal/self-test + shellcheck for scripts/check-matrix-wire-guards.sh.
