+++ src/channels/matrix.rs
@@ -49,7 +49,8 @@ pub(crate) use recovery::{
     MatrixRecoveryCleanupArtifactResult, MatrixRecoveryCleanupArtifactResultState,
     MatrixRecoveryCleanupArtifactRole, MatrixRecoveryCleanupJournal,
     MatrixRecoveryCleanupJournalArtifact, MatrixRecoveryCleanupJournalPhase,
-    MATRIX_RECOVERY_CLEANUP_JOURNAL_VERSION, MATRIX_RECOVERY_KEY_FILE_MAX_BYTES,
+    MatrixRecoveryKeyRotateOutcome, MATRIX_RECOVERY_CLEANUP_JOURNAL_VERSION,
+    MATRIX_RECOVERY_KEY_FILE_MAX_BYTES,
 };
 pub use verification::{
     MatrixSasEmoji, MatrixSasInfo, MatrixVerificationInfo, MatrixVerificationState,
@@ -83,47 +84,6 @@ const MATRIX_UNSUPPORTED_ROOMS_LIMIT: usize = 100;
 /// list is the "recent and small enough to glance at" surface.
 const MATRIX_INBOUND_DLQ_LOST_IDS_CAP: usize = 32;
 pub const MATRIX_STORE_INFO: &[u8] = b"carapace-matrix-store-v1";
-const MATRIX_INBOUND_DLQ_INFO: &[u8] = b"carapace-matrix-inbound-dlq-v1";
-// AAD for the AES-GCM seal of DLQ records. Note: this string lacks the
-// `carapace-` application prefix that `MATRIX_STORE_INFO` and
-// `MATRIX_INBOUND_DLQ_INFO` carry. The prefix omission was the original
-// shape committed and is now locked in: the AAD is part of the GCM
-// authentication tag of every on-disk DLQ record, so changing the
-// constant value is a wire-format break that would invalidate every
-// existing encrypted DLQ. A future v2 envelope can adopt the
-// `carapace-` prefix; until then, this departure is intentional.
-const MATRIX_INBOUND_DLQ_AAD: &[u8] = b"matrix-inbound-dlq-v1";
-/// On-disk envelope version for `MatrixEncryptedInboundDlqRecord`.
-/// Bumping the codec (new field, different encoding, different AAD)
-/// requires bumping this AND the decode-time gate together. Pinning a
-/// named constant keeps writer and reader in sync — two raw `1` literals
-/// at separate sites would let one drift silently and either reject our
-/// own DLQ records or silently mis-decode them with stale params.
-///
-/// v1 (legacy): HKDF-SHA256 over `(passphrase, installation_id)` with
-///   `MATRIX_INBOUND_DLQ_INFO` as the info string. Fast (microseconds
-///   per derivation) — vulnerable to offline brute-force on
-///   `CARAPACE_CONFIG_PASSWORD` if a local attacker has read access to
-///   `state_dir/matrix/inbound_dlq.jsonl` plus
-///   `state_dir/installation_id`.
-/// v2 (current): Argon2id (memory-hard KDF with work factor)
-///   over `(passphrase, installation_id)` via
-///   `crate::crypto::derive_key_argon2id`. The derivation cost
-///   matches the existing config-secret seal layer. Brute-force on
-///   `CARAPACE_CONFIG_PASSWORD` is now memory-bound rather than
-///   HKDF-fast.
-///
-/// Migration: writers always emit v2. Readers accept v1 OR v2 so
-/// existing on-disk records keep decoding through the upgrade —
-/// operators do not need to drain the DLQ before bumping carapace.
-const MATRIX_INBOUND_DLQ_ENVELOPE_VERSION: u8 = 2;
-/// Legacy envelope version. The v1 read path is retained so existing
-/// on-disk records (HKDF-derived keys) continue to decode after
-/// upgrade. Once an operator has fully drained the DLQ, no v1
-/// records remain on disk and the legacy branch is unreachable;
-/// it stays in the source for cross-version compatibility within
-/// the supported upgrade window.
-const MATRIX_INBOUND_DLQ_ENVELOPE_VERSION_LEGACY: u8 = 1;
 const MATRIX_OUTBOUND_QUEUE_CAPACITY: usize = 128;
 /// Maximum inbound Matrix message body size (bytes) before the
 /// runtime drops the event with a warn. A peer in any joined room
@@ -6704,6 +6664,80 @@ fn try_now_millis() -> Result<i64, MatrixError> {
 }
 
 #[cfg(test)]
+fn matrix_rs_fn_body(fn_signature_prefix: &str) -> String {
+    for (label, raw_source) in [
+        (
+            "matrix/verification.rs",
+            include_str!("matrix/verification.rs"),
+        ),
+        ("matrix/recovery.rs", include_str!("matrix/recovery.rs")),
+        (
+            "matrix/inbound_dlq.rs",
+            include_str!("matrix/inbound_dlq.rs"),
+        ),
+        ("matrix.rs", include_str!("matrix.rs")),
+    ] {
+        let source = raw_source.replace("\r\n", "\n");
+        if let Some(fn_start) = source.find(fn_signature_prefix) {
+            let body_offset = source[fn_start..].find("\n}\n").unwrap_or_else(|| {
+                panic!("{fn_signature_prefix} in {label} must have a `\\n}}\\n` closing brace")
+            });
+            return source[fn_start..fn_start + body_offset].to_string();
+        }
+    }
+    panic!("{fn_signature_prefix} must exist in Matrix module sources");
+}
+
+#[cfg(test)]
+fn matrix_test_config(encrypted: bool) -> MatrixConfig {
+    let passphrase = crate::crypto::generate_hex_secret(32).expect("getrandom passphrase");
+    let mut auto_join = MatrixAutoJoinConfig::default();
+    auto_join
+        .allow_users
+        .insert("@alice:example.com".to_string());
+    MatrixConfig {
+        homeserver_url: "https://matrix.example.com".to_string(),
+        user_id: "@cara:example.com".to_string(),
+        access_token: Some(zeroize::Zeroizing::new("token".to_string())),
+        password: None,
+        device_id: Some("DEVICE".to_string()),
+        security: if encrypted {
+            MatrixSecurity::Encrypted {
+                passphrase_source: PassphraseSource::Explicit(
+                    NonEmptyPassphrase::new(&passphrase).expect("passphrase"),
+                ),
+            }
+        } else {
+            MatrixSecurity::Unencrypted
+        },
+        auto_join,
+        legacy_dlq_envelope_policy: MatrixLegacyDlqEnvelopePolicy::Accept,
+    }
+}
+
+#[cfg(test)]
+fn matrix_test_config_with_passphrase(passphrase: &str) -> MatrixConfig {
+    let mut auto_join = MatrixAutoJoinConfig::default();
+    auto_join
+        .allow_users
+        .insert("@alice:example.com".to_string());
+    MatrixConfig {
+        homeserver_url: "https://matrix.example.com".to_string(),
+        user_id: "@cara:example.com".to_string(),
+        access_token: Some(zeroize::Zeroizing::new("token".to_string())),
+        password: None,
+        device_id: Some("DEVICE".to_string()),
+        security: MatrixSecurity::Encrypted {
+            passphrase_source: PassphraseSource::Explicit(
+                NonEmptyPassphrase::new(passphrase).expect("passphrase"),
+            ),
+        },
+        auto_join,
+        legacy_dlq_envelope_policy: MatrixLegacyDlqEnvelopePolicy::Accept,
+    }
+}
+
+#[cfg(test)]
 mod tests {
     use super::inbound_dlq::*;
     use super::recovery::*;
@@ -6772,73 +6806,6 @@ mod tests {
             10,
             &fired_at_zero
         ));
-    }
-
-    /// Pin `matrix_inbound_dlq_line_count` returns `Ok(None)` when
-    /// the file doesn't exist. Append paths call this before opening
-    /// the file; a missing file means "no DLQ entries yet, fine to
-    /// proceed."
-    #[tokio::test]
-    async fn test_matrix_inbound_dlq_line_count_missing_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("missing.jsonl");
-        let result = matrix_inbound_dlq_line_count(&path).await.unwrap();
-        assert_eq!(result, None);
-    }
-
-    /// Pin the byte-floor short-circuit: a file below the
-    /// `CAP_BYTES_FLOOR` heuristic returns `Some(0)` without doing
-    /// any content I/O. This is the hot-path optimization that
-    /// prevents holding dlq_io_lock during full-file reads on the
-    /// common (well-below-cap) case.
-    #[tokio::test]
-    async fn test_matrix_inbound_dlq_line_count_below_floor_returns_zero_sentinel() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("below_floor.jsonl");
-        // Write a small file (well below CAP_BYTES_FLOOR = 10_000 *
-        // 100 = 1 MB). A 4-line file is ~50 bytes total.
-        tokio::fs::write(&path, b"a\nb\nc\nd\n").await.unwrap();
-        let result = matrix_inbound_dlq_line_count(&path).await.unwrap();
-        // Sentinel `Some(0)` short-circuit — the function did NOT
-        // read content. The caller compares `>= MAX_RECORDS`, so 0
-        // is structurally safe.
-        assert_eq!(
-            result,
-            Some(0),
-            "below-floor file must short-circuit to Some(0)"
-        );
-    }
-
-    /// Pin the above-floor full-read path: when the byte size
-    /// crosses the heuristic floor, the function reads the file and
-    /// counts newlines for an exact count (with early-exit at
-    /// MAX_RECORDS).
-    #[tokio::test]
-    async fn test_matrix_inbound_dlq_line_count_above_floor_counts_lines() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("above_floor.jsonl");
-        // Write a file just over CAP_BYTES_FLOOR with a known line
-        // count. CAP_BYTES_FLOOR = 10_000 * 100 = 1_000_000 bytes.
-        // Use 110-byte lines × 10_001 lines = ~1.1 MB, above floor.
-        let line = "x".repeat(109) + "\n"; // 110 bytes
-        let line_count = 10_001;
-        let mut content = String::with_capacity(line.len() * line_count);
-        for _ in 0..line_count {
-            content.push_str(&line);
-        }
-        tokio::fs::write(&path, content.as_bytes()).await.unwrap();
-
-        let result = matrix_inbound_dlq_line_count(&path).await.unwrap();
-        // The function early-exits once `count > MAX_RECORDS`, so
-        // the returned value is `MAX_RECORDS + 1` (10_001) not the
-        // total line count. Either way it triggers the cap branch
-        // in the caller's `>= MAX_RECORDS` check.
-        let count = result.expect("above-floor must produce Some(_)");
-        assert!(
-            count >= MATRIX_INBOUND_DLQ_MAX_RECORDS,
-            "above-floor count must be >= {} for cap branch to fire; got {count}",
-            MATRIX_INBOUND_DLQ_MAX_RECORDS
-        );
     }
 
     /// The inbound-body empty-skip predicate must consider bodies
@@ -7612,28 +7579,6 @@ mod tests {
         }
     }
 
-    fn matrix_test_config_with_passphrase(passphrase: &str) -> MatrixConfig {
-        // Pre-allow the standard test sender; see `matrix_test_config`.
-        let mut auto_join = MatrixAutoJoinConfig::default();
-        auto_join
-            .allow_users
-            .insert("@alice:example.com".to_string());
-        MatrixConfig {
-            homeserver_url: "https://matrix.example.com".to_string(),
-            user_id: "@cara:example.com".to_string(),
-            access_token: Some(zeroize::Zeroizing::new("token".to_string())),
-            password: None,
-            device_id: Some("DEVICE".to_string()),
-            security: MatrixSecurity::Encrypted {
-                passphrase_source: PassphraseSource::Explicit(
-                    NonEmptyPassphrase::new(passphrase).expect("passphrase"),
-                ),
-            },
-            auto_join,
-            legacy_dlq_envelope_policy: MatrixLegacyDlqEnvelopePolicy::Accept,
-        }
-    }
-
     #[derive(Clone)]
     struct FakeInviteSource {
         rooms: Vec<FakeInviteRoom>,
@@ -7737,180 +7682,6 @@ mod tests {
         }
     }
 
-    #[derive(Default)]
-    struct RecordingDlqDispatcher {
-        records: ParkingMutex<Vec<MatrixInboundDlqRecord>>,
-        fail_dispatch: bool,
-    }
-
-    impl RecordingDlqDispatcher {
-        fn failing() -> Self {
-            Self {
-                records: ParkingMutex::new(Vec::new()),
-                fail_dispatch: true,
-            }
-        }
-
-        fn records(&self) -> Vec<MatrixInboundDlqRecord> {
-            self.records.lock().clone()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl MatrixDlqDispatcher for RecordingDlqDispatcher {
-        async fn dispatch(
-            &self,
-            _ws_state: Arc<WsServerState>,
-            _state: Arc<RwLock<MatrixRuntimeState>>,
-            _state_dir: &Path,
-            _config: &MatrixConfig,
-            record: &MatrixInboundDlqRecord,
-        ) -> Result<(), MatrixError> {
-            self.records.lock().push(record.clone());
-            if self.fail_dispatch {
-                Err(MatrixError::SyncFailed(
-                    "scripted DLQ dispatch failure".into(),
-                ))
-            } else {
-                Ok(())
-            }
-        }
-    }
-
-    #[test]
-    fn test_sha256_hasher_zeroizes_on_drop_for_recovery_digests() {
-        fn assert_zeroizes_on_drop<T: zeroize::ZeroizeOnDrop>() {}
-        assert_zeroizes_on_drop::<Sha256>();
-    }
-
-    /// Pin the cap edge: a file at exactly `MATRIX_RECOVERY_KEY_FILE_MAX_BYTES`
-    /// bytes must succeed; a file at cap + 1 must be refused with the
-    /// "exceeds N bytes" error AND no path disclosure. Prevents an
-    /// off-by-one in the post-read `buf.len() > cap` check that would
-    /// either reject the cap-boundary case or accept one byte over.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_read_recovery_key_file_to_string_bounded_at_cap_edges() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let dir = temp.path().join("matrix");
-        std::fs::create_dir_all(&dir).expect("create matrix dir");
-
-        // At cap: succeed.
-        let at_cap = dir.join("recovery_key.at_cap");
-        std::fs::write(
-            &at_cap,
-            vec![b'x'; MATRIX_RECOVERY_KEY_FILE_MAX_BYTES as usize],
-        )
-        .expect("write at-cap file");
-        let result =
-            read_recovery_key_file_to_string_bounded(&at_cap, "Matrix recovery key digest")
-                .await
-                .expect("at-cap read should not error");
-        let bytes = result.expect("at-cap file should yield Some");
-        assert_eq!(bytes.len(), MATRIX_RECOVERY_KEY_FILE_MAX_BYTES as usize);
-
-        // At cap + 1: refuse.
-        let over_cap = dir.join("recovery_key.over_cap");
-        std::fs::write(
-            &over_cap,
-            vec![b'x'; (MATRIX_RECOVERY_KEY_FILE_MAX_BYTES + 1) as usize],
-        )
-        .expect("write over-cap file");
-        let err = read_recovery_key_file_to_string_bounded(&over_cap, "Matrix recovery key digest")
-            .await
-            .expect_err("over-cap read should fail");
-        let msg = err.to_string();
-        assert!(
-            msg.contains(&format!(
-                "exceeds {} bytes",
-                MATRIX_RECOVERY_KEY_FILE_MAX_BYTES
-            )),
-            "over-cap error must surface the cap value: {msg}"
-        );
-        assert!(
-            !msg.contains(&over_cap.display().to_string()),
-            "over-cap error must not expose artifact path: {msg}"
-        );
-    }
-
-    /// Pin the missing-file case: a NotFound returns Ok(None), not Err.
-    /// The daemon's startup probe and rekey recovery rely on this to
-    /// distinguish "no key on disk yet" from "key read failed".
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_read_recovery_key_file_to_string_bounded_missing_returns_none() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let missing = temp.path().join("matrix").join("recovery_key.missing");
-        let result =
-            read_recovery_key_file_to_string_bounded(&missing, "Matrix recovery key digest")
-                .await
-                .expect("missing-file read should not error");
-        assert!(result.is_none(), "missing file must yield None");
-    }
-
-    /// Pin the empty-file case: a 0-byte file returns Some("") which
-    /// the caller (`maybe_restore_recovery_key`) treats as "empty
-    /// after trim" → fail-closed with the operator-actionable
-    /// "recovery key missing" error rather than passing empty bytes
-    /// to the SDK.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_read_recovery_key_file_to_string_bounded_empty_file() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let dir = temp.path().join("matrix");
-        std::fs::create_dir_all(&dir).expect("create matrix dir");
-        let empty = dir.join("recovery_key.empty");
-        std::fs::write(&empty, b"").expect("write empty file");
-
-        let result = read_recovery_key_file_to_string_bounded(&empty, "Matrix recovery key digest")
-            .await
-            .expect("empty-file read should not error");
-        let bytes = result.expect("empty file should yield Some");
-        assert!(bytes.is_empty(), "empty file should yield empty string");
-    }
-
-    /// Pin the non-regular-file refusal: a directory at the path
-    /// must fail with the "not a regular file" error AND not expose
-    /// the path. Companion to `test_recovery_key_digest_read_errors_do_not_expose_paths`
-    /// which exercises the path-disclosure-prevention specifically.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_read_recovery_key_file_to_string_bounded_refuses_directory() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let dir_at_key_path = temp.path().join("matrix").join("recovery_key.dir");
-        std::fs::create_dir_all(&dir_at_key_path).expect("create dir at key path");
-
-        let err = read_recovery_key_file_to_string_bounded(
-            &dir_at_key_path,
-            "Matrix recovery key digest",
-        )
-        .await
-        .expect_err("directory read should fail");
-        let msg = err.to_string();
-        assert!(
-            msg.contains("not a regular file"),
-            "directory error must mention non-regular-file: {msg}"
-        );
-        assert!(
-            !msg.contains(&dir_at_key_path.display().to_string()),
-            "directory error must not expose artifact path: {msg}"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recovery_key_digest_read_errors_do_not_expose_paths() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let path = temp.path().join("matrix").join("recovery_key");
-        std::fs::create_dir_all(&path).expect("create directory at key path");
-
-        let err = recovery_key_file_sha256(&path)
-            .await
-            .expect_err("directory read should fail");
-
-        let message = err.to_string();
-        assert!(message.contains("failed to read Matrix recovery key digest"));
-        assert!(
-            !message.contains(&path.display().to_string()),
-            "recovery-key digest errors must not expose artifact paths: {message}"
-        );
-    }
-
     #[test]
     fn test_encrypted_matrix_state_platform_support_matches_windows_acl_stance() {
         let config = matrix_test_config(true);
@@ -7936,1219 +7707,6 @@ mod tests {
         let config = matrix_test_config(false);
         ensure_encrypted_matrix_state_supported(&config)
             .expect("unencrypted Matrix state must not be blocked by the Windows ACL stance");
-    }
-
-    fn matrix_test_dlq_record() -> MatrixInboundDlqRecord {
-        MatrixInboundDlqRecord {
-            event_id: "$event:example.com".to_string(),
-            room_id: "!room:example.com".to_string(),
-            sender_id: "@alice:example.com".to_string(),
-            text: "encrypted room secret".to_string(),
-            received_at: 1_700_000_000_000,
-        }
-    }
-
-    fn encode_legacy_v1_matrix_inbound_dlq_record_for_test(
-        state_dir: &Path,
-        config: &MatrixConfig,
-        record: &MatrixInboundDlqRecord,
-    ) -> String {
-        let installation_id = read_or_create_installation_id(state_dir).expect("installation_id");
-        let passphrase = resolve_matrix_store_passphrase(state_dir, config)
-            .expect("resolve")
-            .expect("passphrase present");
-        let v1_key = derive_matrix_inbound_dlq_key_v1_from(
-            passphrase.as_bytes(),
-            installation_id.as_bytes(),
-        )
-        .expect("v1 derive");
-
-        let plaintext = serde_json::to_vec(record).expect("serialize record");
-        let blob = crate::crypto::encrypt_aead_blob(&v1_key, &plaintext, MATRIX_INBOUND_DLQ_AAD)
-            .expect("encrypt v1");
-        serde_json::to_string(&MatrixEncryptedInboundDlqRecord {
-            version: MATRIX_INBOUND_DLQ_ENVELOPE_VERSION_LEGACY,
-            nonce: URL_SAFE_NO_PAD.encode(blob.nonce),
-            ciphertext: URL_SAFE_NO_PAD.encode(blob.ciphertext),
-        })
-        .expect("serialize v1 envelope")
-    }
-
-    #[test]
-    fn test_matrix_inbound_dlq_encrypts_records_when_matrix_encrypted() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let record = matrix_test_dlq_record();
-
-        let line = encode_matrix_inbound_dlq_record(temp.path(), &config, &record)
-            .expect("encrypted DLQ line");
-
-        assert!(
-            !line.contains(&record.text),
-            "encrypted Matrix DLQ line must not persist plaintext message body"
-        );
-        assert!(
-            !line.contains(&record.sender_id),
-            "encrypted Matrix DLQ line must not persist plaintext sender"
-        );
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &config, &line)
-            .expect("encrypted DLQ line should decode");
-        assert_eq!(decoded, record);
-    }
-
-    /// Pin the v1 wire shape (HKDF) and verify v2 readers decode it
-    /// correctly. Operators upgrading carapace should NOT need to
-    /// drain the DLQ first — existing v1-encoded records on disk
-    /// must keep decoding through the dual-version branch.
-    #[test]
-    fn test_matrix_inbound_dlq_decodes_legacy_v1_envelope() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let record = matrix_test_dlq_record();
-
-        let v1_line =
-            encode_legacy_v1_matrix_inbound_dlq_record_for_test(temp.path(), &config, &record);
-
-        // v2-deployed reader decodes the v1 line.
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &config, &v1_line)
-            .expect("v1 envelope must decode under v2 reader");
-        assert_eq!(decoded, record);
-    }
-
-    #[test]
-    fn test_matrix_inbound_dlq_refuses_legacy_v1_when_policy_refuse() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let mut config = matrix_test_config(true);
-        config.legacy_dlq_envelope_policy = MatrixLegacyDlqEnvelopePolicy::Refuse;
-        let record = matrix_test_dlq_record();
-
-        let v1_line =
-            encode_legacy_v1_matrix_inbound_dlq_record_for_test(temp.path(), &config, &record);
-
-        let err = decode_matrix_inbound_dlq_record(temp.path(), &config, &v1_line)
-            .expect_err("operator-refuse policy must reject legacy v1 envelopes");
-
-        assert!(matches!(err, MatrixError::LegacyDlqEnvelopeRefused));
-        assert!(
-            err.to_string()
-                .contains("legacy Matrix inbound DLQ v1 envelope refused by policy"),
-            "unexpected error: {err}"
-        );
-        // Post-Batch-79: refused-legacy records are NOT classified as
-        // temporarily-undecodable any more. `policy=Refuse` is the
-        // operator's explicit choice and no toggle makes them
-        // decodable later — so the replay loop routes them to
-        // quarantine (Corrupt) for operator-attended forensic
-        // preservation rather than the live-DLQ "preserved last"
-        // tail-truncation class that silently drops under cap pressure.
-        assert!(
-            !is_temporarily_undecodable_dlq_error(&err),
-            "refused legacy records must NOT be classified as temporarily-undecodable; they belong in quarantine, not the live DLQ tail"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_matrix_inbound_dlq_replay_rewrites_legacy_v1_with_audit() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let record = matrix_test_dlq_record();
-        let v1_line =
-            encode_legacy_v1_matrix_inbound_dlq_record_for_test(temp.path(), &config, &record);
-        let path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&path, format!("{v1_line}\n")).expect("write legacy DLQ");
-
-        let session_root = temp.path().join("sessions-file");
-        std::fs::write(&session_root, b"not a directory").expect("seed session-store blocker");
-        let session_store = Arc::new(crate::sessions::SessionStore::with_base_path(session_root));
-        let ws_state = Arc::new(
-            crate::server::ws::WsServerState::new(crate::server::ws::WsServerConfig::default())
-                .with_session_store(session_store),
-        );
-        let err = replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state)
-            .await
-            .expect_err("dispatch is intentionally unwired, so record remains for retry");
-        assert!(
-            err.to_string()
-                .contains("Matrix inbound DLQ replay still has"),
-            "unexpected replay error: {err}"
-        );
-
-        let rewritten = std::fs::read_to_string(&path).expect("read rewritten DLQ");
-        let envelope: MatrixEncryptedInboundDlqRecord =
-            serde_json::from_str(rewritten.trim()).expect("rewritten encrypted envelope");
-        assert_eq!(
-            envelope.version, MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-            "legacy v1 record must be rewritten as the current DLQ envelope"
-        );
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &config, rewritten.trim())
-            .expect("rewritten v2 line must decode");
-        assert_eq!(decoded, record);
-
-        let audit = std::fs::read_to_string(temp.path().join("audit.jsonl"))
-            .expect("legacy DLQ processing must leave audit evidence");
-        let entry: crate::logging::audit::AuditEntry =
-            serde_json::from_str(audit.lines().next().expect("audit line")).unwrap();
-        assert_eq!(entry.event, "matrix_inbound_dlq_legacy_envelope_processed");
-        assert_eq!(entry.data["from_version"], serde_json::json!(1));
-        assert_eq!(entry.data["current_version"], serde_json::json!(2));
-        assert_eq!(entry.data["record_count"], serde_json::json!(1));
-        assert_eq!(entry.data["reencoded_count"], serde_json::json!(1));
-    }
-
-    /// New writes always emit v2 (Argon2id). A reader can confirm
-    /// the wire shape by parsing the envelope and inspecting the
-    /// `version` field.
-    #[test]
-    fn test_matrix_inbound_dlq_writes_emit_v2_envelope() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let record = matrix_test_dlq_record();
-
-        let line = encode_matrix_inbound_dlq_record(temp.path(), &config, &record)
-            .expect("encrypted DLQ line");
-        let envelope: MatrixEncryptedInboundDlqRecord =
-            serde_json::from_str(&line).expect("parse envelope");
-        assert_eq!(
-            envelope.version, MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-            "new writes must emit the current envelope version (v2 / Argon2id)"
-        );
-        assert_eq!(MATRIX_INBOUND_DLQ_ENVELOPE_VERSION, 2);
-        assert_eq!(MATRIX_INBOUND_DLQ_ENVELOPE_VERSION_LEGACY, 1);
-    }
-
-    #[test]
-    fn test_matrix_inbound_dlq_aad_binds_envelope_version_for_new_records() {
-        let key = zeroize::Zeroizing::new([7u8; crate::crypto::AEAD_KEY_LEN]);
-        let plaintext = b"{\"event_id\":\"$event:example.com\"}";
-        let aad = matrix_inbound_dlq_aad(MATRIX_INBOUND_DLQ_ENVELOPE_VERSION);
-        assert_eq!(
-            aad, b"matrix-inbound-dlq-envelope-v2",
-            "v2 AAD is a released wire-format input and must not drift under version 2"
-        );
-        let blob = crate::crypto::encrypt_aead_blob(&key, plaintext, &aad).expect("encrypt");
-
-        let decoded = decrypt_matrix_inbound_dlq_blob(
-            &key,
-            &blob.nonce,
-            &blob.ciphertext,
-            MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-        )
-        .expect("current version AAD should decrypt");
-        assert_eq!(decoded.as_slice(), plaintext);
-
-        let err = decrypt_matrix_inbound_dlq_blob(
-            &key,
-            &blob.nonce,
-            &blob.ciphertext,
-            MATRIX_INBOUND_DLQ_ENVELOPE_VERSION_LEGACY,
-        )
-        .expect_err("new records must not decrypt under a different envelope version");
-        assert!(
-            err.to_string().contains("decrypt Matrix inbound DLQ"),
-            "expected AAD-bound decrypt failure, got {err}"
-        );
-    }
-
-    #[test]
-    fn test_matrix_inbound_dlq_rejects_legacy_aad_v2_envelope() {
-        let key = zeroize::Zeroizing::new([9u8; crate::crypto::AEAD_KEY_LEN]);
-        let plaintext = b"{\"event_id\":\"$event:example.com\"}";
-        let blob = crate::crypto::encrypt_aead_blob(&key, plaintext, MATRIX_INBOUND_DLQ_AAD)
-            .expect("legacy encrypt");
-
-        let err = decrypt_matrix_inbound_dlq_blob(
-            &key,
-            &blob.nonce,
-            &blob.ciphertext,
-            MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-        )
-        .expect_err("legacy-AAD fallback must be restricted to v1 envelopes");
-
-        assert!(err.to_string().contains("decrypt Matrix inbound DLQ"));
-    }
-
-    #[test]
-    fn test_matrix_encrypted_inbound_dlq_record_rejects_unknown_fields() {
-        let payload = serde_json::json!({
-            "version": MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-            "nonce": "nonce",
-            "ciphertext": "ciphertext",
-            "futureField": true,
-        });
-
-        let result = serde_json::from_value::<MatrixEncryptedInboundDlqRecord>(payload);
-
-        assert!(
-            result.is_err(),
-            "encrypted DLQ envelopes must reject unknown persisted fields"
-        );
-    }
-
-    #[test]
-    fn test_replace_matrix_inbound_dlq_lines_removes_file_when_drained() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(path.parent().unwrap()).expect("create matrix dir");
-        std::fs::write(&path, "record\n").expect("write live dlq");
-
-        replace_matrix_inbound_dlq_lines_blocking(&path, &[]).expect("drain live dlq");
-
-        assert!(
-            !path.exists(),
-            "draining every inbound DLQ record should remove the live file"
-        );
-    }
-
-    /// Cross-version round-trip: encode-as-v2 → decode-via-shared-
-    /// reader should not silently confuse with a v1 record. A v2
-    /// envelope decoded under a hypothetical "v1-only reader" would
-    /// mis-derive the key and AEAD tag mismatch would surface as
-    /// `decrypt Matrix inbound DLQ`. Pin that the v2 reader does
-    /// NOT accidentally decode under v1's KDF.
-    #[test]
-    fn test_matrix_inbound_dlq_v2_record_does_not_decode_with_v1_key() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let record = matrix_test_dlq_record();
-
-        let line =
-            encode_matrix_inbound_dlq_record(temp.path(), &config, &record).expect("v2 line");
-        // Maliciously rewrite the version to v1 so the reader
-        // would attempt v1 (HKDF) decryption against v2 (Argon2id)
-        // ciphertext. AEAD tag mismatch must surface as a decrypt
-        // error, not silent garbage.
-        let mut envelope: MatrixEncryptedInboundDlqRecord =
-            serde_json::from_str(&line).expect("parse v2 envelope");
-        envelope.version = MATRIX_INBOUND_DLQ_ENVELOPE_VERSION_LEGACY;
-        let tampered = serde_json::to_string(&envelope).expect("re-serialize");
-
-        let err = decode_matrix_inbound_dlq_record(temp.path(), &config, &tampered)
-            .expect_err("v2 ciphertext under v1 KDF must fail to decrypt");
-        let msg = err.to_string();
-        assert!(
-            msg.contains("decrypt Matrix inbound DLQ"),
-            "expected AEAD decrypt failure, got: {msg}"
-        );
-    }
-
-    /// Adversary-reachable replay-loop DoS regression. Pre-fix the
-    /// replay loop scanned all on-disk lines for the literal
-    /// substring `"version":1` / `"version":2` regardless of
-    /// `config.encrypted()`. A peer-controlled inbound message body
-    /// (which JSON-encodes inside the plaintext DLQ record's
-    /// `text` field) carrying that literal substring would force
-    /// `ensure_v2`, which fails with `MissingStoreSecret` in
-    /// plaintext config (no passphrase resolvable). Replay then
-    /// aborts phase 1, the dlq_replay streak goes sticky, and the
-    /// channel pins in Error indefinitely. The fix gates the scan
-    /// on `config.encrypted()` so plaintext mode never derives —
-    /// any encrypted-shape lines surface as per-record corrupt
-    /// during decode and get quarantined.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_replay_plaintext_config_with_version_substring_in_body() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-
-        // Append a plaintext record whose body carries the literal
-        // substring `"version":2`. JSON-encodes as a normal string.
-        let record = MatrixInboundDlqRecord {
-            event_id: "$evt:example.com".to_string(),
-            room_id: "!room:example.com".to_string(),
-            sender_id: "@alice:example.com".to_string(),
-            text: r#"discusses "version":2 of the spec"#.to_string(),
-            received_at: 1_700_000_000_000,
-        };
-        append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect("append plaintext record with substring body");
-
-        // Replay must NOT fail with MissingStoreSecret. With the
-        // fix, no derivation runs in plaintext mode regardless of
-        // body bytes. The dispatch will fail (no ws state / no
-        // session machinery wired here), but that failure path is
-        // separate from the DoS regression we're pinning. The pin
-        // is: replay does NOT short-circuit with MissingStoreSecret.
-        let ws_state = Arc::new(crate::server::ws::WsServerState::new(
-            crate::server::ws::WsServerConfig::default(),
-        ));
-        let result = replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state.clone()).await;
-
-        // Even if dispatch fails, the error must NOT be MissingStoreSecret.
-        // Acceptable shapes:
-        //  - Ok(()) (replay succeeded; dispatch wasn't actually invoked)
-        //  - Err(MatrixError::SyncFailed(_)) carrying dispatch failure
-        // NOT acceptable: Err(MatrixError::MissingStoreSecret).
-        if let Err(err) = &result {
-            assert!(
-                !matches!(err, MatrixError::MissingStoreSecret),
-                "plaintext replay must NOT short-circuit with \
-                 MissingStoreSecret on body-substring match: {err:?}"
-            );
-        }
-    }
-
-    /// `MatrixDlqKeys` Debug impl must NOT print the cached AEAD
-    /// key bytes. Hand-rolled per `MatrixInboundDlqRecord` discipline:
-    /// a future contributor adding `#[derive(Debug)]` would silently
-    /// regress AEAD-key-leak protection. The leak only manifests
-    /// when `tracing::debug!(?state, ...)` runs against a populated
-    /// runtime — invisible at compile time, devastating in operator
-    /// logs.
-    #[test]
-    fn test_matrix_dlq_keys_debug_does_not_print_key_bytes() {
-        // Construct a key with a distinctive byte pattern so we
-        // can assert its absence in the Debug output.
-        let keys = MatrixDlqKeys::empty();
-        let pattern: [u8; crate::crypto::AEAD_KEY_LEN] = [0xAB; crate::crypto::AEAD_KEY_LEN];
-        let _ = keys.v2.set(zeroize::Zeroizing::new(pattern));
-
-        let dbg = format!("{keys:?}");
-        // The summary must surface set/unset state.
-        assert!(
-            dbg.contains("<set>"),
-            "Debug must indicate v2 is set: {dbg}"
-        );
-        assert!(
-            dbg.contains("<unset>"),
-            "Debug must indicate v1 is unset: {dbg}"
-        );
-        // The byte pattern must NOT appear (in any format —
-        // decimal, hex, comma-separated array form).
-        assert!(
-            !dbg.contains("171"),
-            "Debug must not print decimal byte values (0xAB = 171): {dbg}"
-        );
-        assert!(
-            !dbg.to_lowercase().contains("ab, ab"),
-            "Debug must not print hex-comma byte values: {dbg}"
-        );
-        assert!(
-            !dbg.contains("[171,"),
-            "Debug must not print array-form byte values: {dbg}"
-        );
-    }
-
-    /// Encrypted-toggle scenario: a daemon previously running with
-    /// `matrix.encrypted=true` left v2 records on disk; a fresh
-    /// daemon start with `matrix.encrypted=false` and an empty
-    /// cache must surface the typed `SyncFailed` error pointing at
-    /// the toggle-back recovery path, NOT panic. Pre-fix the inner
-    /// decode `expect`-panicked here.
-    #[test]
-    fn test_decode_v2_envelope_under_plaintext_config_returns_typed_error() {
-        // Synthesize a v2 envelope (we don't need real AEAD here,
-        // just the envelope shape that triggers the inner's
-        // version-dispatch).
-        let envelope = MatrixEncryptedInboundDlqRecord {
-            version: MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-            nonce: URL_SAFE_NO_PAD.encode([0u8; crate::crypto::AEAD_NONCE_LEN]),
-            ciphertext: URL_SAFE_NO_PAD.encode(b"ciphertext-stub"),
-        };
-        let line = serde_json::to_string(&envelope).expect("serialize envelope");
-
-        // Empty cache + plaintext config = no key derivable. The
-        // inner must return the typed error rather than panic.
-        let plaintext = matrix_test_config(false);
-        let err = decode_matrix_inbound_dlq_record(
-            std::path::Path::new("/nonexistent"),
-            &plaintext,
-            &line,
-        )
-        .expect_err("plaintext config + v2 envelope must surface typed error");
-        let msg = err.to_string();
-        assert!(
-            matches!(
-                err,
-                MatrixError::SyncFailed(_) | MatrixError::MissingStoreSecret
-            ),
-            "expected SyncFailed or MissingStoreSecret, got: {err:?}"
-        );
-        // If SyncFailed, message must point at the toggle-back
-        // recovery path so operators can act.
-        if matches!(err, MatrixError::SyncFailed(_)) {
-            assert!(
-                msg.contains("toggle back to true to drain"),
-                "SyncFailed must point at recovery path: {msg}"
-            );
-        }
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_append_matrix_inbound_dlq_uses_owner_only_file() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let record = matrix_test_dlq_record();
-
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        append_matrix_inbound_dlq(temp.path(), &config, state, &record)
-            .await
-            .expect("append DLQ");
-
-        let path = matrix_inbound_dlq_path(temp.path());
-        assert!(path.exists());
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mode = std::fs::metadata(&path)
-                .expect("metadata")
-                .permissions()
-                .mode()
-                & 0o777;
-            assert_eq!(mode, 0o600);
-        }
-    }
-
-    /// Successful DLQ append must clear a previously sticky
-    /// `inbound_dlq_durability_error`. The pre-fix behavior pinned the
-    /// channel in Error for the daemon's lifetime after a single
-    /// transient append failure even though every later append
-    /// succeeded.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_durability_error_clears_on_successful_append() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let record = matrix_test_dlq_record();
-
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        state
-            .write()
-            .record_inbound_dlq_append_failure("transient EIO".to_string());
-        assert!(state.read().inbound_durability_error_is_sticky());
-
-        append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect("append DLQ");
-
-        assert!(
-            !state.read().inbound_durability_error_is_sticky(),
-            "successful append must clear sticky durability error"
-        );
-    }
-
-    /// Pin the at-cap latch short-circuit: when
-    /// `inbound_dlq_at_cap_since_ms` is fresh (less than 10s ago), an
-    /// append MUST fail-fast with the "latched" failure-mode message
-    /// without touching the on-disk DLQ. The latch was added so a
-    /// sustained inbound-failure flood doesn't pay a full
-    /// read_to_string per record after the cap has already been
-    /// confirmed; this test pins that contract.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_append_matrix_inbound_dlq_short_circuits_when_latch_fresh() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let record = matrix_test_dlq_record();
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-
-        state.write().inbound_dlq_at_cap_since_ms = Some(now_millis());
-        let err = append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect_err("fresh latch must short-circuit append");
-        assert!(
-            matches!(err, MatrixError::SyncFailed(ref msg) if msg.contains("latched")),
-            "latched failure mode must surface as SyncFailed/latched: {err:?}"
-        );
-        let path = matrix_inbound_dlq_path(temp.path());
-        assert!(
-            !path.exists(),
-            "short-circuit must not touch the DLQ file when no file was present"
-        );
-        assert!(
-            state.read().inbound_durability_error_is_sticky(),
-            "short-circuit must still record the per-event drop as a durability event so the \
-             operator-visible sticky-Error signal stays accurate"
-        );
-    }
-
-    /// Companion to the short-circuit test: when the latch is older
-    /// than the TTL (MATRIX_INBOUND_DLQ_AT_CAP_LATCH_TTL_MS = 10s),
-    /// the next append falls through the pre-lock latch check and
-    /// re-confirms cap from disk. Pin by making the latch ancient and
-    /// running an append against an empty DLQ — the append should
-    /// succeed.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_append_matrix_inbound_dlq_falls_through_when_latch_expired() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let record = matrix_test_dlq_record();
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-
-        let ancient = now_millis().saturating_sub(MATRIX_INBOUND_DLQ_AT_CAP_LATCH_TTL_MS + 1_000);
-        state.write().inbound_dlq_at_cap_since_ms = Some(ancient);
-        append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect("expired latch must allow append");
-        let path = matrix_inbound_dlq_path(temp.path());
-        assert!(
-            path.exists(),
-            "fall-through path must commit the append once cap is re-confirmed below limit"
-        );
-    }
-
-    /// Replay of an empty/missing DLQ must clear a sticky durability
-    /// error, matching the append decay path.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_durability_error_clears_on_empty_replay() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        state
-            .write()
-            .record_inbound_dlq_append_failure("transient EIO".to_string());
-        assert!(state.read().inbound_durability_error_is_sticky());
-
-        let ws_state = Arc::new(crate::server::ws::WsServerState::new(
-            crate::server::ws::WsServerConfig::default(),
-        ));
-        replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state.clone())
-            .await
-            .expect("empty replay");
-        assert!(
-            !state.read().inbound_durability_error_is_sticky(),
-            "empty replay tick must decay sticky durability error"
-        );
-    }
-
-    /// A DLQ record whose persisted `event_id` is empty must still be
-    /// rejected, while non-empty raw Matrix IDs with control/display-hostile
-    /// bytes are preserved and replayed with a hash-derived idempotency key.
-    #[test]
-    fn test_decode_matrix_inbound_dlq_record_rejects_empty_event_id() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        // Plaintext encoding so the line is human-readable and we
-        // can hand-craft an "empty event_id" record.
-        let line = serde_json::json!({
-            "eventId": "",
-            "roomId": "!room:example.com",
-            "senderId": "@alice:example.com",
-            "text": "hello",
-            "receivedAt": 1_700_000_000_000_i64,
-        })
-        .to_string();
-        let err = decode_matrix_inbound_dlq_record(temp.path(), &config, &line)
-            .expect_err("empty event_id must be rejected");
-        let msg = err.to_string();
-        assert!(
-            msg.contains("invalid empty event_id"),
-            "expected invalid-empty-event_id error, got {msg}"
-        );
-
-        // Embedded control bytes are no longer rejected: the record keeps the
-        // raw Matrix event_id and replay dedupes via a stable SHA-256 key.
-        let line = serde_json::json!({
-            "eventId": "abc\u{0007}def",
-            "roomId": "!room:example.com",
-            "senderId": "@alice:example.com",
-            "text": "hello",
-            "receivedAt": 1_700_000_000_000_i64,
-        })
-        .to_string();
-        let record = decode_matrix_inbound_dlq_record(temp.path(), &config, &line)
-            .expect("control-byte event_id should decode for hash-idempotent replay");
-        let key = matrix_event_idempotency_key(&record.event_id).expect("hash key");
-        assert!(key.as_str().starts_with("matrix-event-v3-sha256:"));
-    }
-
-    /// A line that fails to decode must NOT be silently dropped on
-    /// the next replay. The current implementation moves corrupt
-    /// lines to an `inbound_dlq.corrupt.jsonl` quarantine sibling so
-    /// the live DLQ can drain (otherwise every replay tick
-    /// re-classifies the same lines as Corrupt and the channel stays
-    /// sticky-Error). Both invariants are pinned here:
-    /// (a) the corrupt line is preserved verbatim in quarantine,
-    /// (b) the live DLQ no longer contains it after replay.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_replay_keeps_corrupt_lines_verbatim() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let path = matrix_inbound_dlq_path(temp.path());
-        let quarantine_path = matrix_inbound_dlq_quarantine_path(temp.path());
-        if let Some(parent) = path.parent() {
-            tokio::fs::create_dir_all(parent).await.expect("dir");
-        }
-        // Pre-seed with a line that won't decode (truncated JSON).
-        let corrupt_line = "{\"eventId\":\"$abc:example.com\",  // truncated\n";
-        tokio::fs::write(&path, corrupt_line)
-            .await
-            .expect("seed corrupt DLQ");
-
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let ws_state = Arc::new(crate::server::ws::WsServerState::new(
-            crate::server::ws::WsServerConfig::default(),
-        ));
-        let err = replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state.clone())
-            .await
-            .expect_err("corrupt-only replay must surface error");
-        assert!(format!("{err}").contains("undecodable"));
-
-        // (a) The corrupt line must be preserved verbatim in the
-        // quarantine file for forensic recovery.
-        let quarantined = tokio::fs::read_to_string(&quarantine_path)
-            .await
-            .expect("read quarantine");
-        assert!(
-            quarantined.contains("$abc:example.com"),
-            "corrupt DLQ line must be quarantined verbatim, got {quarantined:?}"
-        );
-
-        // (b) The live DLQ must no longer contain the corrupt line —
-        // either the file was rewritten empty (then deleted, hence
-        // NotFound), or it exists but contains no records.
-        match tokio::fs::read_to_string(&path).await {
-            Ok(live) => assert!(
-                !live.contains("$abc:example.com"),
-                "corrupt line must be removed from live DLQ after quarantine, got {live:?}"
-            ),
-            Err(err) if err.kind() == std::io::ErrorKind::NotFound => {}
-            Err(err) => panic!("unexpected live DLQ read error: {err}"),
-        }
-    }
-
-    /// `update_verification_record_state` must preserve previously
-    /// captured SAS data when the caller passes `SasUpdate::Preserve`
-    /// — regression test pinning that emoji stay visible across state
-    /// transitions instead of being clobbered on every refresh
-    /// iteration.
-    #[test]
-    fn test_update_verification_record_state_preserves_sas_when_next_is_none() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (info, inserted) = upsert_verification_record(
-            &state,
-            "abc123".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        let flow_id = info.flow_id.clone();
-        let seeded_sas = MatrixSasInfo {
-            emoji: Some(vec![MatrixSasEmoji {
-                symbol: "🐱".to_string(),
-                description: "Cat".to_string(),
-            }]),
-            decimals: Some([1, 2, 3]),
-        };
-        update_verification_record_state(
-            &state,
-            &flow_id,
-            MatrixVerificationState::KeysExchanged,
-            SasUpdate::Set(seeded_sas.clone()),
-        )
-        .expect("seed SAS");
-
-        // Preserve must keep the SAS intact even though the state advances.
-        update_verification_record_state(
-            &state,
-            &flow_id,
-            MatrixVerificationState::Confirmed,
-            SasUpdate::Preserve,
-        )
-        .expect("preserve");
-        let stored = state
-            .read()
-            .verifications
-            .iter()
-            .find(|f| f.flow_id == flow_id)
-            .cloned()
-            .expect("flow exists");
-        assert_eq!(stored.state, MatrixVerificationState::Confirmed);
-        assert_eq!(stored.sas, Some(seeded_sas.clone()));
-
-        // Set replaces.
-        let replacement = MatrixSasInfo {
-            emoji: None,
-            decimals: Some([9, 8, 7]),
-        };
-        update_verification_record_state(
-            &state,
-            &flow_id,
-            MatrixVerificationState::Done,
-            SasUpdate::Set(replacement.clone()),
-        )
-        .expect("set");
-        let stored = state
-            .read()
-            .verifications
-            .iter()
-            .find(|f| f.flow_id == flow_id)
-            .cloned()
-            .expect("flow exists");
-        assert_eq!(stored.sas, Some(replacement));
-    }
-
-    /// `update_verification_record_state` returns `Ok(None)` when the
-    /// requested next state and SAS are equal to the existing record
-    /// (no-op tick) — refresh broadcasts only fire on real change;
-    /// this test pins that contract.
-    #[test]
-    fn test_update_verification_record_state_returns_none_on_noop() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (info, inserted) = upsert_verification_record(
-            &state,
-            "noop-flow".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        let flow_id = info.flow_id.clone();
-        // Same state, SasUpdate::Preserve (so SAS stays None) — must
-        // return Ok(None) so the broadcast caller knows nothing changed.
-        let result = update_verification_record_state(
-            &state,
-            &flow_id,
-            MatrixVerificationState::Requested,
-            SasUpdate::Preserve,
-        )
-        .expect("call");
-        assert!(
-            result.is_none(),
-            "no-op state update must return None so refresh broadcasts only fire on real change"
-        );
-    }
-
-    /// A real state change must return `Ok(Some(record))` so the
-    /// caller can broadcast `matrix.verification.updated` with the
-    /// post-state record.
-    #[test]
-    fn test_update_verification_record_state_returns_some_on_change() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (info, _) = upsert_verification_record(
-            &state,
-            "change-flow".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        let flow_id = info.flow_id.clone();
-        let result = update_verification_record_state(
-            &state,
-            &flow_id,
-            MatrixVerificationState::KeysExchanged,
-            SasUpdate::Preserve,
-        )
-        .expect("call");
-        let updated = result.expect("real change must return Some");
-        assert_eq!(updated.state, MatrixVerificationState::KeysExchanged);
-    }
-
-    /// A SAS-only change with the same state must still report the
-    /// change. Otherwise a refresh that captures fresh SAS data on a
-    /// flow that's been in `KeysExchanged` for a while would fail to
-    /// broadcast and the operator UI would never see the comparison
-    /// emoji.
-    #[test]
-    fn test_update_verification_record_state_returns_some_when_sas_changes() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (info, _) = upsert_verification_record(
-            &state,
-            "sas-flow".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE".to_string()),
-            MatrixVerificationState::KeysExchanged,
-        )
-        .unwrap_applied();
-        let flow_id = info.flow_id.clone();
-        let new_sas = MatrixSasInfo {
-            emoji: None,
-            decimals: Some([1, 2, 3]),
-        };
-        let result = update_verification_record_state(
-            &state,
-            &flow_id,
-            MatrixVerificationState::KeysExchanged,
-            SasUpdate::Set(new_sas.clone()),
-        )
-        .expect("call");
-        let updated = result.expect("SAS change must return Some");
-        assert_eq!(updated.sas, Some(new_sas));
-    }
-
-    /// DLQ replay-success path: a record that dispatches successfully
-    /// must result in the file being removed by `rewrite_matrix_inbound_dlq`
-    /// AND the inbound-failure streak reset on the runtime state.
-    /// Pins the item-93 promised "successful deletion" + "counter
-    /// reset" behaviors. Installs a real session store and activity
-    /// service so dispatch reaches the "no-LLM-provider → Ok" path
-    /// that legitimately produces a successful drain — without that
-    /// setup the test would short-circuit through the not-found branch
-    /// and never exercise the rewrite path.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_replay_drains_succeeding_record_and_resets_inbound_streak() {
-        use crate::channels::activity::ActivityService;
-        use crate::server::ws::WsServerConfig;
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-
-        // Real WsServerState wired with a tempdir-backed session store
-        // and activity service so dispatch_inbound_text_with_options
-        // reaches its non-failing terminal branch (no LLM provider →
-        // returns Ok with run_spawned=false). Without these, dispatch
-        // panics on `state.session_store()`.
-        let cfg = serde_json::json!({});
-        crate::config::clear_cache();
-        crate::config::update_cache(cfg.clone(), cfg.clone());
-        let session_dir = tempfile::tempdir().expect("session tempdir");
-        let session_store = Arc::new(crate::sessions::SessionStore::with_base_path(
-            session_dir.path().to_path_buf(),
-        ));
-        let activity_service = Arc::new(ActivityService::with_limits_for_test(8, 1));
-        let ws_state = Arc::new(
-            crate::server::ws::WsServerState::new(WsServerConfig::default())
-                .with_session_store(session_store)
-                .with_activity_service(activity_service),
-        );
-
-        // Drive the inbound streak to a non-zero count and seed a
-        // sticky durability error so the test can observe both
-        // recoveries.
-        state.write().record_inbound_failure();
-        state.write().record_inbound_failure();
-        state
-            .write()
-            .record_inbound_dlq_append_failure("transient EIO".to_string());
-
-        let record = matrix_test_dlq_record();
-        append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect("append DLQ");
-        let path = matrix_inbound_dlq_path(temp.path());
-        assert!(path.exists(), "DLQ file present before replay");
-
-        replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state.clone())
-            .await
-            .expect("replay must succeed when dispatch returns Ok");
-
-        // Real drain by `rewrite_matrix_inbound_dlq`: the file must
-        // be REMOVED, not just rewritten empty.
-        assert!(
-            !path.exists(),
-            "successfully-replayed DLQ must be drained by rewrite_matrix_inbound_dlq"
-        );
-        assert!(
-            !state.read().inbound_durability_error_is_sticky(),
-            "successful drain must clear the sticky durability error"
-        );
-        assert!(
-            !state.read().inbound_streak_is_sticky(),
-            "successful dispatch must reset the inbound failure streak"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_replay_decodes_encrypted_record_through_fake_dispatcher() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let record = matrix_test_dlq_record();
-        let path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(path.parent().expect("DLQ parent")).expect("create DLQ parent");
-        let line =
-            encode_matrix_inbound_dlq_record(temp.path(), &config, &record).expect("encode DLQ");
-        assert!(
-            line.contains("\"version\":2"),
-            "encrypted replay fixture must use the current v2 DLQ envelope"
-        );
-        std::fs::write(&path, format!("{line}\n")).expect("write DLQ line");
-
-        state
-            .write()
-            .record_inbound_dlq_append_failure("transient EIO".to_string());
-        assert!(
-            state.read().inbound_durability_error_is_sticky(),
-            "test precondition: fake-dispatched replay must clear a planted durability error"
-        );
-
-        let dispatcher = RecordingDlqDispatcher::default();
-        let ws_state = Arc::new(crate::server::ws::WsServerState::new(
-            crate::server::ws::WsServerConfig::default(),
-        ));
-        replay_matrix_inbound_dlq_with_dispatcher(
-            temp.path(),
-            &config,
-            ws_state,
-            state.clone(),
-            &dispatcher,
-        )
-        .await
-        .expect("encrypted DLQ replay must dispatch successfully");
-
-        assert_eq!(
-            dispatcher.records(),
-            vec![record],
-            "fake dispatcher must see the decoded plaintext record from the encrypted on-disk line"
-        );
-        assert!(
-            !path.exists(),
-            "successful fake-dispatched replay must drain the encrypted DLQ file"
-        );
-        assert!(
-            !state.read().inbound_durability_error_is_sticky(),
-            "successful fake-dispatched replay must leave no sticky durability error"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_replay_fake_dispatch_failure_retains_decoded_record() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let record = matrix_test_dlq_record();
-        let path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(path.parent().expect("DLQ parent")).expect("create DLQ parent");
-        let line =
-            encode_matrix_inbound_dlq_record(temp.path(), &config, &record).expect("encode DLQ");
-        std::fs::write(&path, format!("{line}\n")).expect("write DLQ line");
-
-        let dispatcher = RecordingDlqDispatcher::failing();
-        let ws_state = Arc::new(crate::server::ws::WsServerState::new(
-            crate::server::ws::WsServerConfig::default(),
-        ));
-        let err = replay_matrix_inbound_dlq_with_dispatcher(
-            temp.path(),
-            &config,
-            ws_state,
-            state,
-            &dispatcher,
-        )
-        .await
-        .expect_err("scripted dispatch failure must keep the record retryable");
-
-        assert!(
-            err.to_string()
-                .contains("Matrix inbound DLQ replay still has 1 undelivered"),
-            "dispatch failure must propagate the replay-retention summary: {err}"
-        );
-        assert_eq!(
-            dispatcher.records(),
-            vec![record],
-            "fake dispatcher must receive the decoded record before replay retains it"
-        );
-        assert!(
-            path.exists(),
-            "dispatch-failed encrypted record must remain on disk for the next replay tick"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_dlq_replay_quarantines_session_history_corruption() {
-        use crate::channels::activity::ActivityService;
-        use crate::server::ws::WsServerConfig;
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(false);
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let cfg = serde_json::json!({});
-        crate::config::clear_cache();
-        crate::config::update_cache(cfg.clone(), cfg.clone());
-        let session_dir = tempfile::tempdir().expect("session tempdir");
-        let session_store = Arc::new(crate::sessions::SessionStore::with_base_path(
-            session_dir.path().to_path_buf(),
-        ));
-        let activity_service = Arc::new(ActivityService::with_limits_for_test(8, 1));
-        let ws_state = Arc::new(
-            crate::server::ws::WsServerState::new(WsServerConfig::default())
-                .with_session_store(session_store.clone())
-                .with_activity_service(activity_service),
-        );
-
-        let session = session_store
-            .get_or_create_session(
-                "matrix:!room:example.com",
-                crate::sessions::SessionMetadata {
-                    channel: Some(MATRIX_CHANNEL_ID.to_string()),
-                    chat_id: Some("!room:example.com".to_string()),
-                    ..Default::default()
-                },
-            )
-            .expect("create Matrix session");
-        session_store
-            .append_message(crate::sessions::ChatMessage::user(
-                session.id.clone(),
-                "seed",
-            ))
-            .expect("seed history");
-        let history_path = session_dir.path().join(format!("{}.jsonl", session.id));
-        use std::io::Write as _;
-        let mut history = std::fs::OpenOptions::new()
-            .append(true)
-            .open(&history_path)
-            .expect("open history");
-        history.write_all(b"{not-json\n").expect("corrupt history");
-        history.sync_all().expect("sync corrupt history");
-
-        let record = matrix_test_dlq_record();
-        append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect("append DLQ");
-        let dlq_path = matrix_inbound_dlq_path(temp.path());
-        let err = replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state.clone())
-            .await
-            .expect_err("session-history corruption should be reported after quarantine");
-
-        assert!(
-            err.to_string().contains("session-history corruption"),
-            "expected typed session-history corruption in replay error, got {err}"
-        );
-        assert!(
-            !dlq_path.exists(),
-            "permanent session-history corruption should not be retried forever in live DLQ"
-        );
-        let quarantined =
-            tokio::fs::read_to_string(matrix_inbound_dlq_quarantine_path(temp.path()))
-                .await
-                .expect("read quarantine");
-        assert!(
-            quarantined.contains(&record.event_id),
-            "quarantine should retain the raw DLQ record for operator repair"
-        );
-    }
-
-    /// Batch 92: when the quarantine file is at cap and new corrupt /
-    /// refused-legacy lines arrive, the tracing-warn alone is easy to
-    /// lose under flood conditions. A durable audit event must fire so
-    /// the operator's explicit policy decision (which routed the records
-    /// to quarantine in the first place) does not silently lose records
-    /// without a grep-able audit trail.
-    #[tokio::test]
-    async fn test_quarantine_cap_drop_emits_durable_audit_event() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let state_dir = temp.path();
-        let quarantine_path = matrix_inbound_dlq_quarantine_path(state_dir);
-        tokio::fs::create_dir_all(quarantine_path.parent().unwrap())
-            .await
-            .expect("mkdir matrix");
-
-        // Pre-fill the quarantine to AT the cap so even one small line
-        // pushes it over.
-        let filler = vec![b'x'; MATRIX_DLQ_QUARANTINE_MAX_BYTES as usize];
-        tokio::fs::write(&quarantine_path, &filler)
-            .await
-            .expect("seed quarantine to cap");
-
-        let new_lines = vec!["{\"event_id\":\"$cap-drop\"}".to_string()];
-        let result = append_matrix_inbound_dlq_quarantine(state_dir, &new_lines).await;
-        assert!(
-            result.is_ok(),
-            "cap-drop path must return Ok after durable audit succeeds: {result:?}"
-        );
-
-        // Quarantine bytes unchanged — drop was honored.
-        let final_bytes = tokio::fs::read(&quarantine_path)
-            .await
-            .expect("read quarantine after cap-drop");
-        assert_eq!(
-            final_bytes, filler,
-            "quarantine must not have grown past the cap"
-        );
-
-        // Audit log must contain the durable cap-drop record.
-        let audit_path = state_dir.join("audit.jsonl");
-        let audit_contents = tokio::fs::read_to_string(&audit_path)
-            .await
-            .expect("audit.jsonl must exist after cap-drop");
-        assert!(
-            audit_contents.contains("matrix_inbound_dlq_quarantine_cap_dropped"),
-            "audit log missing cap-drop event: {audit_contents}"
-        );
-        assert!(
-            audit_contents.contains("\"dropped_lines\":1"),
-            "audit event must record dropped_lines count: {audit_contents}"
-        );
-    }
-
-    /// Companion to `test_quarantine_cap_drop_emits_durable_audit_event`:
-    /// when the FIRST write batch already exceeds the cap (no pre-
-    /// existing quarantine file), the same durable audit must fire.
-    #[tokio::test]
-    async fn test_quarantine_first_write_oversize_emits_durable_audit_event() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let state_dir = temp.path();
-
-        // Single line larger than the cap.
-        let oversize = "x".repeat(MATRIX_DLQ_QUARANTINE_MAX_BYTES as usize + 64);
-        let new_lines = vec![oversize];
-        let result = append_matrix_inbound_dlq_quarantine(state_dir, &new_lines).await;
-        assert!(
-            result.is_ok(),
-            "first-write cap-drop must succeed after durable audit: {result:?}"
-        );
-
-        let quarantine_path = matrix_inbound_dlq_quarantine_path(state_dir);
-        assert!(
-            !quarantine_path.exists(),
-            "first-write oversize batch must not create the quarantine file"
-        );
-
-        let audit_path = state_dir.join("audit.jsonl");
-        let audit_contents = tokio::fs::read_to_string(&audit_path)
-            .await
-            .expect("audit.jsonl must exist after first-write cap-drop");
-        assert!(
-            audit_contents.contains("matrix_inbound_dlq_quarantine_cap_dropped"),
-            "audit log missing first-write cap-drop event: {audit_contents}"
-        );
-        assert!(
-            audit_contents.contains("\"existing_quarantine_bytes\":0"),
-            "first-write event must record existing_quarantine_bytes=0: {audit_contents}"
-        );
-    }
-
-    /// Pin Batch 66: DLQ replay re-checks the sender allowlist against
-    /// the current config (boot snapshot), not the snapshot at append
-    /// time. An operator who removed a peer from `matrix.autoJoin`
-    /// between the original receive and the next replay tick must see
-    /// the queued message dropped rather than dispatched.
-    #[tokio::test]
-    async fn test_dlq_replay_drops_record_when_sender_no_longer_allowed() {
-        use crate::channels::activity::ActivityService;
-        use crate::server::ws::WsServerConfig;
-        let temp = tempfile::tempdir().expect("tempdir");
-        // Config WITHOUT the test sender — simulates an operator who
-        // removed `@alice:example.com` from auto_join after the DLQ
-        // record was originally appended.
-        let mut config = matrix_test_config(false);
-        config.auto_join = MatrixAutoJoinConfig::default();
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let cfg = serde_json::json!({});
-        crate::config::clear_cache();
-        crate::config::update_cache(cfg.clone(), cfg.clone());
-        let session_dir = tempfile::tempdir().expect("session tempdir");
-        let session_store = Arc::new(crate::sessions::SessionStore::with_base_path(
-            session_dir.path().to_path_buf(),
-        ));
-        let activity_service = Arc::new(ActivityService::with_limits_for_test(8, 1));
-        let ws_state = Arc::new(
-            crate::server::ws::WsServerState::new(WsServerConfig::default())
-                .with_session_store(session_store.clone())
-                .with_activity_service(activity_service),
-        );
-
-        let record = matrix_test_dlq_record();
-        append_matrix_inbound_dlq(temp.path(), &config, state.clone(), &record)
-            .await
-            .expect("append DLQ");
-        let dlq_path = matrix_inbound_dlq_path(temp.path());
-        assert!(
-            dlq_path.exists(),
-            "DLQ record should exist on disk before replay"
-        );
-
-        replay_matrix_inbound_dlq(temp.path(), &config, ws_state, state.clone())
-            .await
-            .expect("replay must succeed: dropped records are not errors");
-
-        assert!(
-            !dlq_path.exists(),
-            "DLQ file must be removed (or empty) — the now-disallowed record was dropped, \
-             not left as un-dispatchable backlog occupying the cap"
-        );
     }
 
     /// Test fixture: build a `ChannelRegistry` with the Matrix channel
@@ -9690,63 +8248,6 @@ mod tests {
         );
     }
 
-    /// Pin the JSON shape of `MatrixVerificationInfo` against the
-    /// schema declared in `tests/golden/ws/events.json`. If the
-    /// runtime adds/removes a public field on the verification info,
-    /// the golden schema must be updated in lockstep — this test
-    /// catches drift before WS clients see it.
-    #[test]
-    fn test_matrix_verification_info_serializes_required_fields() {
-        let info = MatrixVerificationInfo {
-            flow_id: "flow-1".to_string(),
-            protocol_flow_id: "txn-1".to_string(),
-            raw_protocol_flow_id: "txn-1".to_string(),
-            user_id: "@alice:example.com".to_string(),
-            device_id: Some("DEVICE".to_string()),
-            state: MatrixVerificationState::Requested,
-            sas: None,
-            created_at: 1,
-            updated_at: 2,
-        };
-        let json = serde_json::to_value(&info).expect("serialize");
-        // Required fields per the WS event schema.
-        for field in [
-            "flowId",
-            "protocolFlowId",
-            "userId",
-            "state",
-            "createdAt",
-            "updatedAt",
-        ] {
-            assert!(
-                json.get(field).is_some(),
-                "MatrixVerificationInfo must serialize {field}"
-            );
-        }
-        // State must serialize to one of the documented values.
-        let state_str = json
-            .get("state")
-            .and_then(|v| v.as_str())
-            .expect("state is string");
-        let allowed = [
-            "created",
-            "requested",
-            "ready",
-            "transitioned",
-            "started",
-            "accepted",
-            "keys_exchanged",
-            "confirmed",
-            "done",
-            "cancelled",
-            "mismatched",
-        ];
-        assert!(
-            allowed.contains(&state_str),
-            "MatrixVerificationState '{state_str}' is not in the documented enum"
-        );
-    }
-
     #[test]
     fn test_failure_streak_record_failure_returns_count() {
         let mut streak = FailureStreak::new(3);
@@ -10027,66 +8528,6 @@ mod tests {
         assert!(
             !msg.contains("!room6 inspect failed"),
             "non-preview entries should not appear in the message: {msg}"
-        );
-    }
-
-    /// Wave-decode helper: cap-clamp truncation must classify
-    /// each tail record as either decoded (collect event_id) or
-    /// undecodable (count toward decode_failures). Real-cap
-    /// (10000 records) is impractical in tests; helper extraction
-    /// lets us pin the accounting on a small fixture.
-    #[test]
-    fn test_collect_dropped_event_ids_classifies_decoded_vs_undecodable() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-
-        // Three real records (all v2 / Argon2id) + two undecodable
-        // lines (random JSON envelopes — wrong ciphertext).
-        let real_records: Vec<MatrixInboundDlqRecord> = (0..3)
-            .map(|i| MatrixInboundDlqRecord {
-                event_id: format!("$evt-{i}:example.com"),
-                room_id: "!room:example.com".to_string(),
-                sender_id: "@alice:example.com".to_string(),
-                text: format!("body {i}"),
-                received_at: 1_700_000_000_000 + i as i64,
-            })
-            .collect();
-        let real_lines: Vec<String> = real_records
-            .iter()
-            .map(|r| {
-                encode_matrix_inbound_dlq_record(temp.path(), &config, r)
-                    .expect("encode real record")
-            })
-            .collect();
-        // Synthesize an envelope-shaped line with garbage
-        // ciphertext — decodes to AEAD failure.
-        let garbage_line = serde_json::to_string(&MatrixEncryptedInboundDlqRecord {
-            version: MATRIX_INBOUND_DLQ_ENVELOPE_VERSION,
-            nonce: URL_SAFE_NO_PAD.encode([0u8; crate::crypto::AEAD_NONCE_LEN]),
-            ciphertext: URL_SAFE_NO_PAD.encode(b"random-junk-not-real-ciphertext"),
-        })
-        .expect("serialize");
-        let mut tail = real_lines.clone();
-        tail.push(garbage_line.clone());
-        tail.push(garbage_line);
-
-        let keys = MatrixDlqKeys::empty();
-        keys.ensure_v2(temp.path(), &config).expect("derive v2");
-        let (dropped_ids, decode_failures) =
-            collect_dropped_event_ids_from_tail(&tail, Some(&keys));
-
-        assert_eq!(
-            decode_failures, 2,
-            "garbage envelopes must surface as decode_failures"
-        );
-        assert_eq!(
-            dropped_ids,
-            vec![
-                "$evt-0:example.com".to_string(),
-                "$evt-1:example.com".to_string(),
-                "$evt-2:example.com".to_string(),
-            ],
-            "decoded event_ids preserved in tail order"
         );
     }
 
@@ -10420,128 +8861,6 @@ mod tests {
         );
     }
 
-    /// Pinned vector for the inbound-DLQ HKDF derivation. Drift here
-    /// is a silent wire-format break: an operator upgrading carapace
-    /// while a non-empty `inbound_dlq.jsonl` is on disk would lose
-    /// access to those records (decryption would yield gibberish or
-    /// AEAD-tag failures, and the records would land in the
-    /// quarantine sibling). Pin against a known input so any rotation
-    /// of `MATRIX_INBOUND_DLQ_INFO`, salt formula, or HKDF-construction
-    /// detail trips this test before shipping.
-    #[test]
-    fn test_pinned_matrix_inbound_dlq_key_vector() {
-        let key = derive_matrix_inbound_dlq_key_from(
-            b"correct horse battery staple",
-            b"installation-00000000-0000-0000-0000-000000000000",
-        )
-        .unwrap();
-        assert_eq!(
-            hex::encode(key),
-            "4f17d4dc2615c81e3e552fe15374de09634d883e4b7835a1d43a04676f5a0ff7"
-        );
-    }
-
-    #[test]
-    fn test_pinned_matrix_inbound_dlq_v2_key_vector() {
-        // Inputs are intentional pinned test fixtures — see v1 vector above.
-        // Routed through `Vec` rather than literal byte slices so static
-        // analysis doesn't misread the pin as a real password/salt.
-        let passphrase: Vec<u8> = Vec::from(&b"correct horse battery staple"[..]);
-        let installation: Vec<u8> =
-            Vec::from(&b"installation-00000000-0000-0000-0000-000000000000"[..]);
-        let key =
-            derive_matrix_inbound_dlq_key_v2_from(passphrase.as_slice(), installation.as_slice())
-                .unwrap();
-        assert_eq!(
-            hex::encode(key),
-            "92a4336c637a2792d43ce389686fb958777d557c27a79c287bac9d8350b12c78"
-        );
-    }
-
-    /// v2 derivation (Argon2id) must be deterministic over
-    /// `(passphrase, installation_id)` so that the encode-time and
-    /// decode-time keys match across daemon restarts. A pinned
-    /// expected hex value would be fragile against future Argon2id
-    /// parameter rotations (memory / iterations / lanes), which are
-    /// signaled by an envelope-version bump rather than a derivation
-    /// drift, so this test asserts the determinism property
-    /// directly: the same inputs produce byte-identical keys, and a
-    /// different input produces a different key.
-    #[test]
-    fn test_v2_argon2id_derivation_is_deterministic() {
-        // Generate non-secret fixtures at test time so this test
-        // doesn't carry hardcoded byte literals into a derivation
-        // function — CodeQL flags hardcoded inputs into crypto
-        // APIs as a code-smell. The determinism property holds
-        // regardless of the specific bytes; we only need
-        // (passphrase, salt) inputs that round-trip stably and
-        // a second pair that differs in passphrase or salt.
-        let mut passphrase = [0u8; 32];
-        let mut salt = [0u8; 32];
-        getrandom::fill(&mut passphrase).expect("getrandom passphrase");
-        getrandom::fill(&mut salt).expect("getrandom salt");
-
-        let key_a = derive_matrix_inbound_dlq_key_v2_from(&passphrase, &salt).expect("v2 derive a");
-        let key_b = derive_matrix_inbound_dlq_key_v2_from(&passphrase, &salt).expect("v2 derive b");
-        assert_eq!(*key_a, *key_b, "Argon2id derivation must be deterministic");
-
-        let mut other_passphrase = [0u8; 32];
-        getrandom::fill(&mut other_passphrase).expect("getrandom other passphrase");
-        let key_c =
-            derive_matrix_inbound_dlq_key_v2_from(&other_passphrase, &salt).expect("v2 derive c");
-        assert_ne!(
-            *key_a, *key_c,
-            "Argon2id keys must differ for different passphrases"
-        );
-
-        let mut other_salt = [0u8; 32];
-        getrandom::fill(&mut other_salt).expect("getrandom other salt");
-        let key_d =
-            derive_matrix_inbound_dlq_key_v2_from(&passphrase, &other_salt).expect("v2 derive d");
-        assert_ne!(
-            *key_a, *key_d,
-            "Argon2id keys must differ for different installation_ids"
-        );
-
-        // v1 (HKDF) and v2 (Argon2id) MUST produce different keys
-        // for the same inputs — sharing a derivation would defeat
-        // the dual-version envelope (a v1 reader could decode v2
-        // ciphertext or vice versa).
-        let v1 = derive_matrix_inbound_dlq_key_v1_from(&passphrase, &salt).expect("v1 derive");
-        assert_ne!(
-            *key_a, *v1,
-            "v1 (HKDF) and v2 (Argon2id) must derive distinct keys for the same inputs"
-        );
-    }
-
-    /// `MatrixDlqKeys::ensure_v2` must cache the derivation result.
-    /// A second call with the same inputs returns the SAME borrowed
-    /// reference (pointer-equal). A future refactor that always
-    /// re-derives (e.g. drops the OnceLock fast-path) would silently
-    /// regress to ~100ms per call, defeating the daemon-lifetime
-    /// cache. The byte-equality check would still pass under that
-    /// regression, but pointer equality cannot.
-    #[test]
-    fn test_matrix_dlq_keys_ensure_v2_cache_idempotence() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let config = matrix_test_config(true);
-        let keys = MatrixDlqKeys::empty();
-        let first = keys.ensure_v2(temp.path(), &config).expect("v2 derive");
-        // Capture the pointer of the first borrow; immediately
-        // release the borrow so we can re-call.
-        let first_ptr = first.as_ptr();
-        let second = keys
-            .ensure_v2(temp.path(), &config)
-            .expect("v2 derive cached");
-        let second_ptr = second.as_ptr();
-        assert_eq!(
-            first_ptr, second_ptr,
-            "ensure_v2 must return the same OnceLock-backed reference \
-             on a second call (cache hit); a regression that re-derives \
-             would have a different pointer"
-        );
-    }
-
     /// Pin: per-path Forbidden classification. `matrix_send_terminal_error`
     /// must route Forbidden to `SendTerminal` (room-level
     /// rejection), NOT `AuthTokenRevoked` (which would mislead
@@ -10577,96 +8896,6 @@ mod tests {
         );
     }
 
-    /// Pin: `upsert_verification_record` stores BOTH the sanitized
-    /// `protocol_flow_id` (operator-display surface) and the raw
-    /// `raw_protocol_flow_id` (SDK lookup key). Sanitization is
-    /// non-bijective; passing the sanitized form to
-    /// `client.encryption().get_verification_*` would fail to
-    /// resolve any flow whose original id contained codepoints
-    /// stripped by sanitize. `apply_verification_action` MUST use
-    /// the raw form for SDK lookups.
-    #[test]
-    fn test_upsert_verification_record_preserves_raw_protocol_flow_id() {
-        let runtime_state = Arc::new(parking_lot::RwLock::new(MatrixRuntimeState::default()));
-        // Inject a hostile-shape protocol flow id with a zero-width
-        // joiner. Sanitize strips ZWJs (U+200D); the raw must be
-        // preserved exactly so SDK lookup matches.
-        let raw_flow = "txn-\u{200d}-abc";
-        let sanitized_flow = "txn--abc";
-        let (info, _inserted) = upsert_verification_record(
-            &runtime_state,
-            raw_flow.to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert_eq!(
-            info.raw_protocol_flow_id, raw_flow,
-            "raw_protocol_flow_id must preserve the original bytes for SDK lookup"
-        );
-        assert_eq!(
-            info.protocol_flow_id, sanitized_flow,
-            "protocol_flow_id must be sanitized for operator-display surfaces"
-        );
-        // Wire serialization MUST omit raw_protocol_flow_id (it
-        // would defeat the whole point: operator scripts decoding the
-        // wire JSON would see un-sanitized bytes).
-        let json = serde_json::to_value(&info).expect("serialize");
-        assert!(
-            json.get("rawProtocolFlowId").is_none() && json.get("raw_protocol_flow_id").is_none(),
-            "raw_protocol_flow_id must NOT serialize to wire JSON"
-        );
-    }
-
-    /// Pin: `apply_verification_action` resolves SDK lookups via the
-    /// `raw_protocol_flow_id` field, never the sanitized form.
-    /// Static-analysis pin against the function body: the let-binding
-    /// must be `flow.raw_protocol_flow_id.clone()` and `protocol_flow_id`
-    /// must NOT be re-bound to `flow.protocol_flow_id` (the sanitized
-    /// form) anywhere in the SDK-call section.
-    #[test]
-    fn test_apply_verification_action_uses_raw_for_sdk_lookup() {
-        let body = matrix_rs_fn_body("async fn apply_verification_action");
-        let body = body.as_str();
-        assert!(
-            body.contains("let protocol_flow_id = flow.raw_protocol_flow_id.clone();"),
-            "apply_verification_action must clone raw_protocol_flow_id for SDK lookup"
-        );
-        assert!(
-            !body.contains("let protocol_flow_id = flow.protocol_flow_id.clone();"),
-            "apply_verification_action must NOT re-bind protocol_flow_id to the \
-             sanitized field; sanitize is non-bijective and SDK lookup would fail"
-        );
-    }
-
-    /// Pin: `refresh_verification_records` also resolves SDK
-    /// lookups via the raw protocol flow id. A ZWSP / bidi /
-    /// control-stripped flow can sit in the daemon list long enough
-    /// for the refresh worker to update it; using the sanitized
-    /// display field there would make refresh mark the flow stale
-    /// while `apply_verification_action` still finds it.
-    #[test]
-    fn test_refresh_verification_records_uses_raw_for_sdk_lookup() {
-        let body = matrix_rs_fn_body("async fn refresh_verification_records");
-        let body = body.as_str();
-        assert!(
-            body.contains(
-                "get_verification_request(&parsed_user_id, &record.raw_protocol_flow_id)"
-            ),
-            "refresh_verification_records must use raw_protocol_flow_id for request lookup"
-        );
-        assert!(
-            body.contains("get_verification(&parsed_user_id, &record.raw_protocol_flow_id)"),
-            "refresh_verification_records must use raw_protocol_flow_id for SAS lookup"
-        );
-        assert!(
-            !body.contains("get_verification_request(&parsed_user_id, &record.protocol_flow_id)")
-                && !body.contains("get_verification(&parsed_user_id, &record.protocol_flow_id)"),
-            "refresh_verification_records must not use the sanitized protocol_flow_id for SDK lookups"
-        );
-    }
-
     /// Pin: room-message verification requests must use the same
     /// sender trust gate as to-device requests before they consume
     /// verification-record capacity.
@@ -10690,64 +8919,6 @@ mod tests {
             body.contains("room-message verification request dropped"),
             "untrusted room-message verification requests must drop before upsert"
         );
-    }
-
-    /// Pin: `append_matrix_inbound_dlq` reaches the v2 key through
-    /// the daemon-lifetime `MatrixDlqKeys` cache, never the
-    /// standalone `derive_matrix_inbound_dlq_key_v2_from`.
-    /// Bypassing the cache regresses concurrent inbound dispatch
-    /// failures to a fresh ~100ms Argon2id derivation per record
-    /// while holding `dlq_io_lock` — byte-equivalence pins don't
-    /// catch this; only call-site routing does.
-    #[test]
-    fn test_append_matrix_inbound_dlq_routes_through_cache() {
-        // Disambiguate from `append_matrix_inbound_dlq_quarantine`,
-        // which has a different prefix-match in the source.
-        let body = matrix_rs_fn_body("async fn append_matrix_inbound_dlq(");
-        let body = body.as_str();
-
-        assert!(
-            body.contains("state.read().dlq_keys()"),
-            "append_matrix_inbound_dlq must fetch the cache via \
-             state.read().dlq_keys() — direct derivation defeats \
-             the daemon-lifetime fast-path"
-        );
-        assert!(
-            body.contains("dlq_keys.ensure_v2(state_dir, config)"),
-            "append_matrix_inbound_dlq must obtain the v2 key via \
-             dlq_keys.ensure_v2(state_dir, config)"
-        );
-        assert!(
-            !body.contains("derive_matrix_inbound_dlq_key_v2_from"),
-            "append_matrix_inbound_dlq must NOT call the standalone \
-             derive helper directly — that path bypasses the cache"
-        );
-    }
-
-    /// `MatrixDlqKeys::ensure_v2` first-derivation-failure must NOT
-    /// poison the OnceLock — a subsequent successful call after the
-    /// operator fixes their config must populate the slot. Otherwise
-    /// transient `MissingStoreSecret` (env var unset for one call)
-    /// would permanently wedge the cache for the daemon's lifetime.
-    #[test]
-    fn test_matrix_dlq_keys_ensure_v2_retries_after_failure() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        // Plaintext config: ensure_v2 fails with MissingStoreSecret.
-        let plain = matrix_test_config(false);
-        let keys = MatrixDlqKeys::empty();
-        let err = keys
-            .ensure_v2(temp.path(), &plain)
-            .expect_err("ensure_v2 must fail with no passphrase");
-        assert!(matches!(err, MatrixError::MissingStoreSecret));
-        // OnceLock must remain empty so the next call can retry.
-        assert!(keys.v2().is_none(), "failure must not poison the OnceLock");
-
-        // Re-attempt with a passphrase available — must succeed.
-        let encrypted = matrix_test_config(true);
-        let _ = keys
-            .ensure_v2(temp.path(), &encrypted)
-            .expect("ensure_v2 must succeed after config is fixed");
-        assert!(keys.v2().is_some());
     }
 
     /// `summarize_failures` boundary cases. Pinned directly — the
@@ -11536,161 +9707,6 @@ mod tests {
     }
 
     #[test]
-    fn test_verification_record_upsert_and_prune() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (first, inserted) = upsert_verification_record(
-            &state,
-            "protocol-flow-1".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE1".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        assert_eq!(first.protocol_flow_id, "protocol-flow-1");
-        assert_eq!(first.state, MatrixVerificationState::Requested);
-
-        let (updated, inserted) = upsert_verification_record(
-            &state,
-            "protocol-flow-1".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE2".to_string()),
-            MatrixVerificationState::Ready,
-        )
-        .unwrap_applied();
-        assert!(!inserted);
-        assert_eq!(updated.device_id.as_deref(), Some("DEVICE2"));
-        assert_eq!(updated.state, MatrixVerificationState::Ready);
-
-        state.write().verifications[0].updated_at =
-            now_millis() - MATRIX_VERIFICATION_RECORD_TTL.as_millis() as i64 - 1;
-        prune_verification_records(&state);
-        assert!(state.read().verifications.is_empty());
-    }
-
-    #[test]
-    fn test_matrix_verification_info_wire_shape_is_camel_case_and_raw_safe() {
-        let info = MatrixVerificationInfo {
-            flow_id: "mvr_test".to_string(),
-            protocol_flow_id: "txn-safe".to_string(),
-            raw_protocol_flow_id: "txn-\u{200b}-raw".to_string(),
-            user_id: "@alice:example.com".to_string(),
-            device_id: Some("DEVICE".to_string()),
-            state: MatrixVerificationState::KeysExchanged,
-            sas: Some(MatrixSasInfo {
-                emoji: Some(vec![MatrixSasEmoji {
-                    symbol: "🐱".to_string(),
-                    description: "Cat".to_string(),
-                }]),
-                decimals: Some([1234, 5678, 9012]),
-            }),
-            created_at: 10,
-            updated_at: 20,
-        };
-
-        let json = serde_json::to_value(&info).expect("serialize verification info");
-        assert_eq!(json["flowId"], "mvr_test");
-        assert_eq!(json["protocolFlowId"], "txn-safe");
-        assert!(json.get("rawProtocolFlowId").is_none());
-        assert!(json.get("raw_protocol_flow_id").is_none());
-        assert_eq!(json["userId"], "@alice:example.com");
-        assert_eq!(json["deviceId"], "DEVICE");
-        assert_eq!(json["state"], "keys_exchanged");
-        assert_eq!(json["createdAt"], 10);
-        assert_eq!(json["updatedAt"], 20);
-        assert_eq!(
-            json.pointer("/sas/decimals"),
-            Some(&json!([1234, 5678, 9012]))
-        );
-        assert_eq!(json.pointer("/sas/emoji/0/symbol"), Some(&json!("🐱")));
-        assert_eq!(
-            json.pointer("/sas/emoji/0/description"),
-            Some(&json!("Cat"))
-        );
-    }
-
-    #[test]
-    fn test_matrix_sas_info_wire_shape_omits_absent_optional_fields() {
-        let sas = MatrixSasInfo {
-            emoji: None,
-            decimals: Some([1, 2, 3]),
-        };
-        let json = serde_json::to_value(&sas).expect("serialize sas info");
-        assert_eq!(json, json!({ "decimals": [1, 2, 3] }));
-
-        let emoji = MatrixSasEmoji {
-            symbol: "7".to_string(),
-            description: "Seven".to_string(),
-        };
-        let json = serde_json::to_value(&emoji).expect("serialize sas emoji");
-        assert_eq!(json, json!({ "symbol": "7", "description": "Seven" }));
-    }
-
-    #[test]
-    fn test_verification_control_id_includes_user_and_protocol_flow() {
-        assert_eq!(
-            matrix_verification_control_id("@alice:example.com", "shared-protocol-flow"),
-            matrix_verification_control_id("@alice:example.com", "shared-protocol-flow"),
-            "daemon control ids must be deterministic for the same raw peer flow"
-        );
-        assert!(
-            matrix_verification_control_id("@alice:example.com", "shared-protocol-flow")
-                .starts_with("mvr_"),
-            "daemon control ids must live in the Matrix verification request namespace"
-        );
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (alice, inserted) = upsert_verification_record(
-            &state,
-            "shared-protocol-flow".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE1".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        let (bob, inserted) = upsert_verification_record(
-            &state,
-            "shared-protocol-flow".to_string(),
-            "@bob:example.com".to_string(),
-            Some("DEVICE2".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        assert_ne!(alice.flow_id, bob.flow_id);
-        assert_eq!(state.read().verifications.len(), 2);
-
-        let sanitized_collision = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let (raw_with_zwsp, inserted) = upsert_verification_record(
-            &sanitized_collision,
-            "flow-\u{200b}-id".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE1".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        let (raw_without_zwsp, inserted) = upsert_verification_record(
-            &sanitized_collision,
-            "flow--id".to_string(),
-            "@alice:example.com".to_string(),
-            Some("DEVICE1".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        assert_eq!(
-            raw_with_zwsp.protocol_flow_id,
-            raw_without_zwsp.protocol_flow_id
-        );
-        assert_ne!(
-            raw_with_zwsp.flow_id, raw_without_zwsp.flow_id,
-            "peer protocol ids that sanitize to the same display id must not collide in daemon control ids"
-        );
-        assert_eq!(sanitized_collision.read().verifications.len(), 2);
-    }
-
-    #[test]
     fn test_send_text_queue_full_returns_retryable_delivery_result() {
         let (tx, _rx) = mpsc::channel::<MatrixCommand>(1);
         let (reply_tx, _reply_rx) = std_mpsc::sync_channel(1);
@@ -12232,94 +10248,6 @@ mod tests {
         );
     }
 
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recovery_key_file_has_secret_bytes_for_nonempty_content() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("recovery_key");
-        std::fs::write(&path, b"EsT8 Pgxc Fake Recovery Key\n").expect("write");
-        assert!(recovery_key_file_has_secret_bytes(&path)
-            .await
-            .expect("probe must succeed"));
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recovery_key_file_has_secret_bytes_for_empty_file() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("recovery_key");
-        std::fs::write(&path, b"").expect("write zero-byte");
-        assert!(
-            !recovery_key_file_has_secret_bytes(&path)
-                .await
-                .expect("probe must succeed"),
-            "zero-byte recovery key file must not count as having secret bytes — \
-             the stale-marker cleanup branch must keep the marker for operator inspection"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recovery_key_file_has_secret_bytes_for_whitespace_only() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("recovery_key");
-        std::fs::write(&path, b"   \n\t\r\n").expect("write whitespace");
-        assert!(
-            !recovery_key_file_has_secret_bytes(&path)
-                .await
-                .expect("probe must succeed"),
-            "whitespace-only recovery key file must not count as having secret bytes"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_cleanup_stale_recovery_minting_marker_removes_marker_when_key_has_content() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let key_path = dir.path().join("recovery_key");
-        let marker_path = dir.path().join("recovery_key.minting");
-        std::fs::write(&key_path, b"EsT8 Pgxc Fake Recovery Key\n").expect("write key");
-        std::fs::write(&marker_path, b"recovery-minting-in-progress\n").expect("write marker");
-
-        cleanup_stale_recovery_minting_marker(&key_path, &marker_path).await;
-
-        assert!(!marker_path.exists(), "marker should have been cleaned up");
-        assert!(key_path.exists(), "key file must remain untouched");
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_cleanup_stale_recovery_minting_marker_preserves_marker_when_key_is_empty() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let key_path = dir.path().join("recovery_key");
-        let marker_path = dir.path().join("recovery_key.minting");
-        std::fs::write(&key_path, b"").expect("write zero-byte key");
-        std::fs::write(&marker_path, b"recovery-minting-in-progress\n").expect("write marker");
-
-        cleanup_stale_recovery_minting_marker(&key_path, &marker_path).await;
-
-        assert!(
-            marker_path.exists(),
-            "marker MUST survive an empty key file — it is the only breadcrumb to the \
-             orphaned server-side mint and dropping it would strand the recovery state"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_cleanup_stale_recovery_minting_marker_is_infallible_when_key_unreadable() {
-        // Even if the key probe errors (e.g., transient I/O issue,
-        // missing path), the cleanup must not panic or propagate a
-        // failure — a healthy daemon must not refuse startup just
-        // because a stale marker happens to be unverifiable.
-        let dir = tempfile::tempdir().expect("tempdir");
-        let key_path = dir.path().join("recovery_key");
-        let marker_path = dir.path().join("recovery_key.minting");
-        // Deliberately do NOT create key_path so the probe surfaces a
-        // typed I/O error.
-        std::fs::write(&marker_path, b"recovery-minting-in-progress\n").expect("write marker");
-
-        cleanup_stale_recovery_minting_marker(&key_path, &marker_path).await;
-
-        // The marker should remain — we don't take destructive action
-        // when we cannot confirm the key file's contents.
-        assert!(marker_path.exists(), "marker MUST survive a probe error");
-    }
-
     /// Regression for R58 H-AC1: when the SDK's `send_matrix_text`
     /// has just returned Ok AND a terminal-cause cancel fires in the
     /// same poll round, the send-task `tokio::select!` must NOT
@@ -12581,18 +10509,6 @@ mod tests {
     }
 
     #[test]
-    fn test_verification_state_serde_round_trip_and_terminal() {
-        let encoded = serde_json::to_value(MatrixVerificationState::KeysExchanged).unwrap();
-        assert_eq!(encoded, json!("keys_exchanged"));
-        let decoded: MatrixVerificationState = serde_json::from_value(encoded).unwrap();
-        assert_eq!(decoded, MatrixVerificationState::KeysExchanged);
-        assert!(MatrixVerificationState::Done.is_terminal());
-        assert!(MatrixVerificationState::Cancelled.is_terminal());
-        assert!(MatrixVerificationState::Mismatched.is_terminal());
-        assert!(!MatrixVerificationState::Ready.is_terminal());
-    }
-
-    #[test]
     fn test_backoff_sequence_caps_at_sixty_seconds() {
         let mut backoff = MatrixBackoff::default();
         let values: Vec<u64> = (0..9).map(|_| backoff.next_delay(None).as_secs()).collect();
@@ -12602,61 +10518,6 @@ mod tests {
         assert_eq!(
             backoff.next_delay(Some(Duration::from_secs(120))),
             Duration::from_secs(120)
-        );
-    }
-
-    /// Pin: `replay_matrix_inbound_dlq` cap-clamp wires the SUFFIX
-    /// (records being dropped) through `collect_dropped_event_ids_from_tail`
-    /// → `record_inbound_dlq_lost_event_ids` → undecodable-count
-    /// surfacing → `truncate`. End-to-end is impractical (10K
-    /// records) and lowering the cap via `#[cfg(test)]` weakens
-    /// the production guard. The two failure modes this catches:
-    /// inverting the slice index (KEPT prefix vs DROPPED suffix)
-    /// and ordering the truncate before the tail-decode (suffix
-    /// becomes empty before decoding).
-    #[test]
-    fn test_replay_matrix_inbound_dlq_cap_clamp_wiring_pinned() {
-        let body = matrix_rs_fn_body("async fn replay_matrix_inbound_dlq_with_dispatcher");
-        let body = body.as_str();
-
-        // The tail must be sliced from the cap (not to it). A
-        // `..MATRIX_INBOUND_DLQ_MAX_RECORDS` slice would describe
-        // the KEPT prefix, not the dropped suffix — silently
-        // logging the wrong event IDs.
-        assert!(
-            body.contains("merged_lines[MATRIX_INBOUND_DLQ_MAX_RECORDS..]"),
-            "cap-clamp must decode the SUFFIX (records being dropped), not \
-             the prefix (records kept) — `merged_lines[MATRIX_INBOUND_DLQ_MAX_RECORDS..]`"
-        );
-        assert!(
-            body.contains("collect_dropped_event_ids_from_tail"),
-            "cap-clamp must call collect_dropped_event_ids_from_tail to \
-             decode the suffix"
-        );
-        assert!(
-            body.contains("record_inbound_dlq_lost_event_ids"),
-            "cap-clamp must surface the dropped event IDs for operator \
-             forensics via record_inbound_dlq_lost_event_ids"
-        );
-        assert!(
-            body.contains("inbound_dlq_undecodable_lost_count"),
-            "cap-clamp must surface undecodable-but-lost count separately \
-             from event IDs (decode_failures > 0 case)"
-        );
-        // Truncation MUST happen after the tail decode. We pin
-        // ordering by checking that `truncate(MATRIX_INBOUND_DLQ_MAX_RECORDS)`
-        // appears after the call to `collect_dropped_event_ids_from_tail`
-        // in source order.
-        let decode_pos = body
-            .find("collect_dropped_event_ids_from_tail")
-            .expect("decode-call must exist");
-        let truncate_pos = body
-            .find("merged_lines.truncate(MATRIX_INBOUND_DLQ_MAX_RECORDS)")
-            .expect("truncate-call must exist");
-        assert!(
-            truncate_pos > decode_pos,
-            "truncate must run AFTER tail-decode; otherwise the suffix \
-             being decoded is empty and dropped_ids is silently zero"
         );
     }
 
@@ -13550,330 +11411,6 @@ mod tests {
         assert!(decode_raw_device_id_hex("").is_err());
     }
 
-    /// Verification-flow eviction must be terminal-first. Otherwise
-    /// a peer flooding fresh protocol_flow_ids fills the cap with
-    /// non-terminal records and evicts the operator's pending flow
-    /// at index 0 — denying the operator's verification UX.
-    #[test]
-    fn test_upsert_verification_record_evicts_terminal_first_at_cap() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        // Fill the cap with 1 OPERATOR-PENDING flow (non-terminal)
-        // followed by (CAP-1) PEER-INITIATED flows in mixed states,
-        // including some terminal so we can verify terminal-first.
-        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
-        // First record: a non-terminal flow at index 0 (the
-        // operator's). It must NOT be evicted on the next overflow.
-        let (_, inserted) = upsert_verification_record(
-            &state,
-            "operator-flow".to_string(),
-            "@operator:example.com".to_string(),
-            Some("OPERDEV".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        // Add cap-2 more non-terminals; record at index 1 will be a
-        // marker we can later spot when it's NOT evicted.
-        for i in 0..(cap - 2) {
-            upsert_verification_record(
-                &state,
-                format!("peer-flow-{i}"),
-                format!("@peer{i}:example.com"),
-                Some(format!("PEERDEV{i}")),
-                MatrixVerificationState::Requested,
-            )
-            .unwrap_applied();
-        }
-        // Add one TERMINAL record at the end (the SECOND-to-last,
-        // since we add one more terminal to ensure terminal eviction).
-        upsert_verification_record(
-            &state,
-            "old-terminal".to_string(),
-            "@cancelled:example.com".to_string(),
-            Some("DEVTERM".to_string()),
-            MatrixVerificationState::Cancelled,
-        )
-        .unwrap_applied();
-        // Cap is now reached. Insert one more — this triggers
-        // eviction. The OLDEST-TERMINAL is "old-terminal"; should be
-        // dropped. Operator's flow stays.
-        let (_, inserted) = upsert_verification_record(
-            &state,
-            "trigger-evict".to_string(),
-            "@new:example.com".to_string(),
-            Some("DEVNEW".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        assert!(inserted);
-        let guard = state.read();
-        let flow_ids: Vec<&str> = guard
-            .verifications
-            .iter()
-            .map(|f| f.protocol_flow_id.as_str())
-            .collect();
-        assert!(
-            flow_ids.contains(&"operator-flow"),
-            "operator's pending flow MUST NOT be evicted under peer flood"
-        );
-        assert!(
-            !flow_ids.contains(&"old-terminal"),
-            "oldest terminal record MUST be evicted before any non-terminal"
-        );
-    }
-
-    /// Multi-call cumulative cap-eviction with a terminal-rich
-    /// stream. As long as terminal records exist in the buffer at
-    /// eviction time, the operator's non-terminal flow at index 0
-    /// must NEVER be evicted. The all-non-terminal case is the
-    /// documented backstop scenario where TTL pruning protects
-    /// the flow; this test explicitly avoids that scenario by
-    /// keeping the buffer terminal-rich (every insert is terminal
-    /// after the operator's flow goes in).
-    #[test]
-    fn test_upsert_verification_record_multi_call_cumulative_eviction_preserves_operator_flow() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
-
-        upsert_verification_record(
-            &state,
-            "operator-flow".to_string(),
-            "@operator:example.com".to_string(),
-            Some("OPERDEV".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-
-        // Drive 3*cap upserts of TERMINAL records (peer-initiated
-        // verifications that immediately resolve to Cancelled).
-        // Each insert above cap triggers eviction; with only
-        // terminals available, the eviction policy drops the
-        // oldest terminal and never touches the operator's
-        // non-terminal at index 0.
-        for i in 0..(3 * cap) {
-            upsert_verification_record(
-                &state,
-                format!("peer-flow-{i}"),
-                format!("@peer{i}:example.com"),
-                Some(format!("DEV{i}")),
-                MatrixVerificationState::Cancelled,
-            )
-            .unwrap_applied();
-        }
-
-        let guard = state.read();
-        assert!(
-            guard.verifications.len() <= cap,
-            "verification cap must hold under sustained upserts: got {} > {cap}",
-            guard.verifications.len(),
-        );
-        let has_operator = guard
-            .verifications
-            .iter()
-            .any(|f| f.protocol_flow_id == "operator-flow");
-        assert!(
-            has_operator,
-            "operator's pending flow at index 0 must survive cumulative terminal-rich \
-             cap eviction — terminal-first eviction policy must protect index 0"
-        );
-    }
-
-    /// All-terminal cap-eviction: when every record in the buffer
-    /// is terminal, eviction drops the oldest terminal (which is
-    /// the oldest record overall — same outcome). Pin that the
-    /// eviction does NOT panic and that the buffer stays bounded.
-    #[test]
-    fn test_upsert_verification_record_all_terminal_at_cap_drops_oldest() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
-        for i in 0..(cap + 5) {
-            upsert_verification_record(
-                &state,
-                format!("flow-{i}"),
-                format!("@user{i}:example.com"),
-                Some(format!("DEV{i}")),
-                MatrixVerificationState::Cancelled,
-            )
-            .unwrap_applied();
-        }
-        let guard = state.read();
-        assert_eq!(guard.verifications.len(), cap);
-        // Oldest 5 records (flow-0 through flow-4) should be
-        // evicted; flow-5 onward should remain.
-        for i in 0..5 {
-            let id = format!("flow-{i}");
-            assert!(
-                !guard.verifications.iter().any(|f| f.protocol_flow_id == id),
-                "flow-{i} (oldest) should have been evicted"
-            );
-        }
-        for i in 5..(cap + 5) {
-            let id = format!("flow-{i}");
-            assert!(
-                guard.verifications.iter().any(|f| f.protocol_flow_id == id),
-                "flow-{i} (recent) should still be in the buffer"
-            );
-        }
-    }
-
-    #[test]
-    fn test_upsert_verification_record_refuses_when_all_records_are_active() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
-        for i in 0..cap {
-            upsert_verification_record(
-                &state,
-                format!("active-flow-{i}"),
-                format!("@user{i}:example.com"),
-                Some(format!("DEV{i}")),
-                MatrixVerificationState::KeysExchanged,
-            )
-            .unwrap_applied();
-        }
-
-        let result = upsert_verification_record(
-            &state,
-            "new-peer-flow".to_string(),
-            "@new:example.com".to_string(),
-            Some("NEWDEV".to_string()),
-            MatrixVerificationState::Requested,
-        );
-
-        assert!(matches!(result, VerificationRecordUpsert::RejectedAtCap));
-        let guard = state.read();
-        assert_eq!(guard.verifications.len(), cap);
-        assert!(
-            guard
-                .verifications
-                .iter()
-                .any(|f| f.protocol_flow_id == "active-flow-0"),
-            "active ceremonies must not be evicted to admit a fresh peer request"
-        );
-        assert!(
-            !guard
-                .verifications
-                .iter()
-                .any(|f| f.protocol_flow_id == "new-peer-flow"),
-            "fresh peer request must not be recorded when every capped slot is active"
-        );
-    }
-
-    /// Exactly-at-cap: no eviction happens on the boundary insert.
-    /// The cap is `MATRIX_VERIFICATION_RECORDS_MAX` items; the
-    /// `cap+1`-th insert triggers the first eviction. Pin that
-    /// exactly-cap is steady-state, no flapping.
-    #[test]
-    fn test_upsert_verification_record_at_cap_does_not_evict() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
-        for i in 0..cap {
-            upsert_verification_record(
-                &state,
-                format!("flow-{i}"),
-                format!("@user{i}:example.com"),
-                Some(format!("DEV{i}")),
-                MatrixVerificationState::Requested,
-            )
-            .unwrap_applied();
-        }
-        let guard = state.read();
-        assert_eq!(
-            guard.verifications.len(),
-            cap,
-            "exactly-cap insert count must produce exactly-cap buffer length"
-        );
-        // All flows must be present (no evictions yet).
-        for i in 0..cap {
-            let id = format!("flow-{i}");
-            assert!(
-                guard.verifications.iter().any(|f| f.protocol_flow_id == id),
-                "flow-{i} must still be in the buffer at exactly-cap"
-            );
-        }
-    }
-
-    /// Verify-action against a flow already in `Cancelled`/`Done`/
-    /// `Mismatched` must surface `MatrixError::VerificationCancelled`,
-    /// NOT a generic Verification error or a state-transition success.
-    /// The CLI routes 410 Gone on this variant, signaling the operator
-    /// to start a new flow rather than retry. This test pins the
-    /// classifier's terminal-state guard at the helper level by
-    /// stamping a verification record into runtime state and reading
-    /// back the post-cancel snapshot.
-    #[test]
-    fn test_upsert_verification_record_preserves_terminal_state_after_cancel() {
-        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
-        // Initial flow: Requested.
-        upsert_verification_record(
-            &state,
-            "flow-1".to_string(),
-            "@peer:example.com".to_string(),
-            Some("DEVPEER".to_string()),
-            MatrixVerificationState::Requested,
-        )
-        .unwrap_applied();
-        // Update to Cancelled (terminal). State machine must store
-        // the terminal state so a later Confirm can see it.
-        let (rec, _) = upsert_verification_record(
-            &state,
-            "flow-1".to_string(),
-            "@peer:example.com".to_string(),
-            Some("DEVPEER".to_string()),
-            MatrixVerificationState::Cancelled,
-        )
-        .unwrap_applied();
-        assert_eq!(rec.state, MatrixVerificationState::Cancelled);
-
-        // The exact `apply_verification_action` plumbing requires
-        // an SDK `VerificationRequest` handle; the testable surface
-        // is the state lookup that `apply_verification_action`
-        // would consult. Pin that the cancelled record is
-        // discoverable by `protocol_flow_id` so the action path
-        // can return `VerificationCancelled`.
-        let guard = state.read();
-        let found = guard
-            .verifications
-            .iter()
-            .find(|f| f.protocol_flow_id == "flow-1")
-            .expect("flow-1 must exist");
-        assert!(found.state.is_terminal());
-        assert_eq!(found.state, MatrixVerificationState::Cancelled);
-    }
-
-    #[test]
-    fn test_verification_terminal_guard_rejects_accept_and_confirm_but_allows_cancel() {
-        for terminal_state in [
-            MatrixVerificationState::Cancelled,
-            MatrixVerificationState::Done,
-            MatrixVerificationState::Mismatched,
-        ] {
-            for action in [
-                MatrixVerificationAction::Accept,
-                MatrixVerificationAction::Confirm { matches: true },
-            ] {
-                let err =
-                    guard_verification_action_terminal_state("flow-1", &action, &terminal_state)
-                        .expect_err(
-                        "accept/confirm against terminal flow must be rejected before SDK lookup",
-                    );
-                assert!(matches!(
-                    err,
-                    MatrixError::VerificationCancelled {
-                        flow_id,
-                        state,
-                    } if flow_id == "flow-1" && state == terminal_state
-                ));
-            }
-
-            guard_verification_action_terminal_state(
-                "flow-1",
-                &MatrixVerificationAction::Cancel,
-                &terminal_state,
-            )
-            .expect("cancel against terminal flow is idempotent");
-        }
-    }
-
     /// `MatrixError::InterruptedRekey`'s Display starts with a
     /// stable, operator-greppable prefix. Operator runbooks and the
     /// CLI's typed-arm routing are designed around the prefix shape;
@@ -13890,780 +11427,6 @@ mod tests {
             msg.starts_with("Matrix store rekey interrupted:"),
             "InterruptedRekey Display prefix must remain stable for operator runbooks: got `{msg}`"
         );
-    }
-
-    #[test]
-    fn test_recovery_key_rotation_marker_json_wire_shape_is_pinned() {
-        let marker = RecoveryKeyRotationMarker {
-            stage: RecoveryKeyRotationMarkerStage::PendingKeyWritten,
-            key_sha256: Some("new-digest".to_string()),
-            previous_key_sha256: Some("previous-digest".to_string()),
-            updated_at_ms: 1234,
-            legacy_text_marker: false,
-        };
-
-        let value = serde_json::to_value(&marker).unwrap();
-        assert_eq!(value["stage"], "pending_key_written");
-        assert_eq!(value["keySha256"], "new-digest");
-        assert_eq!(value["previousKeySha256"], "previous-digest");
-        assert_eq!(value["updatedAtMs"], 1234);
-        assert!(
-            value.get("legacyTextMarker").is_none(),
-            "in-memory legacy sentinel must not leak into persisted marker JSON"
-        );
-
-        let legacy_json = serde_json::json!({
-            "stage": "pending_key_written",
-            "keySha256": "new-digest",
-            "updatedAtMs": 1234
-        });
-        let parsed: RecoveryKeyRotationMarker = serde_json::from_value(legacy_json).unwrap();
-        assert_eq!(
-            parsed.stage,
-            RecoveryKeyRotationMarkerStage::PendingKeyWritten
-        );
-        assert_eq!(parsed.key_sha256.as_deref(), Some("new-digest"));
-        assert_eq!(parsed.previous_key_sha256, None);
-        assert!(
-            !parsed.legacy_text_marker,
-            "typed JSON marker missing previousKeySha256 is not the legacy text sentinel"
-        );
-    }
-
-    #[test]
-    fn test_recovery_key_pending_refusal_audit_payload_is_typed_and_redacted() {
-        let marker = RecoveryKeyRotationMarker {
-            stage: RecoveryKeyRotationMarkerStage::FinalKeyReplaced,
-            key_sha256: Some("new-digest-secret".to_string()),
-            previous_key_sha256: Some("old-digest-secret".to_string()),
-            updated_at_ms: 1234,
-            legacy_text_marker: false,
-        };
-
-        let event = recovery_pending_refusal_event(
-            &marker,
-            crate::logging::audit::MatrixRecoveryKeyPromotionRefusalReason::FinalStagePendingPresent,
-            Some("old-digest-secret"),
-            Some("pending-digest-secret"),
-        );
-        let value = serde_json::to_value(&event).unwrap();
-        assert_eq!(
-            value["type"],
-            serde_json::json!("matrix_recovery_key_pending_promotion_refused")
-        );
-        assert_eq!(value["marker_stage"], "final_key_replaced");
-        assert_eq!(value["reason"], "final_stage_pending_present");
-        assert_eq!(value["current_key"], "matches_previous_key");
-        assert_eq!(value["pending_key"], "mismatch");
-        let serialized = serde_json::to_string(&value).unwrap();
-        assert!(!serialized.contains("digest-secret"));
-        assert!(!serialized.contains("recovery_key.pending"));
-        assert!(!serialized.contains("recovery_key.rotating"));
-        assert!(!serialized.contains('/'));
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recover_interrupted_recovery_key_rotation_clears_completed_marker() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let key = "recovery-key-after-rotation";
-        std::fs::write(&key_path, format!("{key}\n")).expect("write final key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::FinalKeyReplaced,
-            Some(recovery_key_sha256(key)),
-            Some(recovery_key_sha256("recovery-key-before-rotation")),
-        )
-        .await
-        .expect("write completed marker");
-
-        recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect("completed marker should be cleared");
-
-        assert!(
-            !marker_path.exists(),
-            "post-replacement recovery-key rotation marker should be removed"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{key}\n")
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_started_recovery_key_rotation_without_pending_clears_when_current_matches_marker()
-    {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let restored_key = "operator-restored-previous-key";
-        std::fs::write(&key_path, format!("{restored_key}\n")).expect("write restored key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::Started,
-            None,
-            Some(recovery_key_sha256(restored_key)),
-        )
-        .await
-        .expect("write started marker");
-
-        recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect("started marker with no pending key and matching current key should clear");
-
-        assert!(
-            !marker_path.exists(),
-            "restore cleanup crash marker should be cleared"
-        );
-        assert!(
-            !pending_path.exists(),
-            "restore cleanup crash state must not recreate pending key"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{restored_key}\n")
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recover_interrupted_recovery_key_rotation_refuses_started_cleanup_journal() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let journal_path = matrix_recovery_cleanup_journal_path(temp.path());
-        std::fs::create_dir_all(pending_path.parent().unwrap()).expect("create matrix dir");
-        std::fs::write(&pending_path, "pending-recovery-secret\n").expect("write pending key");
-        let journal = MatrixRecoveryCleanupJournal {
-            version: MATRIX_RECOVERY_CLEANUP_JOURNAL_VERSION,
-            phase: MatrixRecoveryCleanupJournalPhase::Started,
-            artifacts: vec![MatrixRecoveryCleanupJournalArtifact {
-                role: MatrixRecoveryCleanupArtifactRole::PendingKey,
-                path: "matrix/recovery_key.pending".to_string(),
-                expected_provenance: "stale_after_operator_restore".to_string(),
-                result: MatrixRecoveryCleanupArtifactResult {
-                    state: MatrixRecoveryCleanupArtifactResultState::Pending,
-                    error_kind: None,
-                },
-            }],
-        };
-        std::fs::write(&journal_path, serde_json::to_vec(&journal).unwrap())
-            .expect("write cleanup journal");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("started cleanup journal must fail closed");
-
-        assert!(
-            err.to_string().contains("cleanup journal"),
-            "unexpected cleanup journal refusal: {err}"
-        );
-        assert!(
-            pending_path.exists(),
-            "startup must not trust or remove pending material from a started cleanup journal"
-        );
-        assert!(journal_path.exists());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recover_interrupted_recovery_key_rotation_clears_completed_cleanup_journal() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let journal_path = matrix_recovery_cleanup_journal_path(temp.path());
-        std::fs::create_dir_all(journal_path.parent().unwrap()).expect("create matrix dir");
-        let journal = MatrixRecoveryCleanupJournal {
-            version: MATRIX_RECOVERY_CLEANUP_JOURNAL_VERSION,
-            phase: MatrixRecoveryCleanupJournalPhase::Completed,
-            artifacts: vec![MatrixRecoveryCleanupJournalArtifact {
-                role: MatrixRecoveryCleanupArtifactRole::PendingKey,
-                path: "matrix/recovery_key.pending".to_string(),
-                expected_provenance: "stale_after_operator_restore".to_string(),
-                result: MatrixRecoveryCleanupArtifactResult {
-                    state: MatrixRecoveryCleanupArtifactResultState::Removed,
-                    error_kind: None,
-                },
-            }],
-        };
-        std::fs::write(&journal_path, serde_json::to_vec(&journal).unwrap())
-            .expect("write cleanup journal");
-
-        recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect("completed cleanup journal should be cleared");
-
-        assert!(
-            !journal_path.exists(),
-            "completed cleanup journal should be removed before startup repair continues"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_load_recovery_rotation_marker_times_out_wedged_read() {
-        let marker_path = Path::new("matrix/recovery_key.rotating");
-        let state_dir = Path::new(".");
-        let err = load_recovery_rotation_marker_with_timeout(
-            marker_path,
-            state_dir,
-            std::time::Duration::ZERO,
-            std::future::pending::<std::io::Result<Vec<u8>>>(),
-        )
-        .await
-        .expect_err("wedged marker read must time out");
-
-        assert!(
-            err.to_string()
-                .contains("timed out reading Matrix recovery-key rotation marker"),
-            "unexpected timeout error: {err}"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recovery_artifact_cleanup_failure_is_returned() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(&marker_path).expect("create marker directory");
-
-        let err = remove_recovery_marker_with_log(&marker_path)
-            .await
-            .expect_err("directory marker cleanup must fail");
-
-        assert!(
-            err.to_string()
-                .contains("failed to remove Matrix recovery marker"),
-            "cleanup failure must be operator-visible, got {err}"
-        );
-        assert!(
-            marker_path.exists(),
-            "failed cleanup must not report success or hide the artifact"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recover_interrupted_recovery_key_rotation_promotes_pending_over_old_current_key()
-    {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let old_key = "recovery-key-before-rotation";
-        let new_key = "recovery-key-after-rotation";
-        std::fs::write(&key_path, format!("{old_key}\n")).expect("write old key");
-        std::fs::write(&pending_path, format!("{new_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::PendingKeyWritten,
-            Some(recovery_key_sha256(new_key)),
-            Some(recovery_key_sha256(old_key)),
-        )
-        .await
-        .expect("write pending marker");
-
-        recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect("pending key with stale old current key should recover");
-
-        assert!(
-            !marker_path.exists(),
-            "stale recovery-key rotation marker should be removed"
-        );
-        assert!(!pending_path.exists(), "pending key should be promoted");
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{new_key}\n")
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_pending_key_written_refuses_pending_when_current_key_missing() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let old_key = "recovery-key-before-rotation";
-        let new_key = "recovery-key-after-rotation";
-        std::fs::write(&pending_path, format!("{new_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::PendingKeyWritten,
-            Some(recovery_key_sha256(new_key)),
-            Some(recovery_key_sha256(old_key)),
-        )
-        .await
-        .expect("write pending marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("pending-key-written marker must fail closed when current key is missing");
-
-        assert!(
-            err.to_string().contains("current key is missing"),
-            "unexpected recovery error: {err}"
-        );
-        assert!(
-            marker_path.exists(),
-            "marker must remain for operator repair"
-        );
-        assert!(pending_path.exists(), "pending key must remain untouched");
-        assert!(
-            !key_path.exists(),
-            "missing current key must not be recreated"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_corrupt_typed_recovery_key_rotation_marker_fails_closed_without_leaking_material()
-    {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        std::fs::write(&key_path, "current-recovery-secret\n").expect("write current key");
-        std::fs::write(&pending_path, "pending-recovery-secret\n").expect("write pending key");
-        std::fs::write(
-            &marker_path,
-            br#"{"stage":"pending_key_written","keySha256":"#,
-        )
-        .expect("write corrupt typed marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("corrupt typed marker must fail closed");
-        let message = err.to_string();
-        let temp_path = temp.path().to_string_lossy().to_string();
-
-        assert!(message.contains("typed recovery-key rotation marker is malformed"));
-        assert!(!message.contains(&temp_path));
-        assert!(!message.contains("current-recovery-secret"));
-        assert!(!message.contains("pending-recovery-secret"));
-        assert!(!message.contains("keySha256"));
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            "current-recovery-secret\n"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&pending_path).unwrap(),
-            "pending-recovery-secret\n"
-        );
-        assert!(
-            marker_path.exists(),
-            "corrupt marker must remain for explicit operator inspection"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_bom_prefixed_typed_recovery_key_rotation_marker_is_corrupt_typed() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        std::fs::write(&key_path, "current-recovery-secret\n").expect("write current key");
-        std::fs::write(&pending_path, "pending-recovery-secret\n").expect("write pending key");
-        std::fs::write(
-            &marker_path,
-            b"\xEF\xBB\xBF{\"stage\":\"pending_key_written\",\"keySha256\":",
-        )
-        .expect("write bom-prefixed corrupt typed marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("BOM-prefixed typed marker must fail closed as typed corruption");
-        assert!(
-            err.to_string()
-                .contains("typed recovery-key rotation marker is malformed"),
-            "BOM-prefixed typed marker must not be classified as unknown legacy: {err}"
-        );
-        assert!(
-            marker_path.exists(),
-            "corrupt marker must remain for explicit operator inspection"
-        );
-    }
-
-    /// Batch 95: the rename helper for recovery-key promotion must
-    /// refuse when the source file's bytes changed between the
-    /// caller's validation and our rename. A same-uid attacker who
-    /// races the dirent swap should NOT be able to commit unverified
-    /// bytes into `recovery_key`.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_replace_owner_only_secret_file_refuses_on_digest_mismatch() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let src = temp.path().join("src");
-        let dst = temp.path().join("dst");
-        std::fs::write(&src, "expected-bytes\n").expect("write src");
-        // Caller's hash (legitimate validation step).
-        let expected_digest = recovery_key_sha256("expected-bytes");
-        // Simulate a TOCTOU swap: between caller's hash and the
-        // rename helper's re-hash, an attacker rewrote the file at
-        // `src`.
-        std::fs::write(&src, "attacker-bytes\n").expect("swap src bytes");
-        let err = replace_owner_only_secret_file(&src, &dst, &expected_digest)
-            .await
-            .expect_err("helper must refuse rename when source digest changed");
-        assert!(
-            err.contains("digest changed"),
-            "expected digest-mismatch refusal, got: {err}"
-        );
-        assert!(
-            !dst.exists(),
-            "rename must not commit attacker bytes when digest mismatch is detected"
-        );
-        assert!(src.exists(), "src dirent should still be present");
-    }
-
-    /// Batch 95 companion: happy path — when the source bytes do
-    /// match the expected digest, the rename completes normally.
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_replace_owner_only_secret_file_succeeds_on_digest_match() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let src = temp.path().join("src");
-        let dst = temp.path().join("dst");
-        let content = "expected-bytes";
-        std::fs::write(&src, format!("{content}\n")).expect("write src");
-        let expected_digest = recovery_key_sha256(content);
-        replace_owner_only_secret_file(&src, &dst, &expected_digest)
-            .await
-            .expect("rename must succeed when src digest matches expected");
-        assert!(dst.exists(), "rename should have committed dst");
-        assert!(!src.exists(), "src should have been moved");
-        assert_eq!(
-            std::fs::read_to_string(&dst).expect("read dst"),
-            format!("{content}\n")
-        );
-    }
-
-    /// Batch 113: `replace_owner_only_secret_file`'s re-read buffer is
-    /// `Zeroizing<Vec<u8>>` with pre-allocated capacity so no realloc
-    /// happens during `read_to_end`. Verify the read path produces
-    /// the same successful outcome regardless of file size up to the
-    /// recovery-key cap. (We can't directly inspect the freed heap
-    /// from a unit test, but a regression on the realloc-free
-    /// allocation contract would show as a panic / OOM at the cap
-    /// boundary — this test exercises that boundary.)
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_replace_owner_only_secret_file_handles_max_size_no_realloc() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let src = temp.path().join("src");
-        let dst = temp.path().join("dst");
-        // Fill the source file to exactly the recovery-key cap (4 KiB).
-        let max_bytes_usize: usize = MATRIX_RECOVERY_KEY_FILE_MAX_BYTES as usize;
-        let content_bytes = vec![b'x'; max_bytes_usize];
-        std::fs::write(&src, &content_bytes).expect("write src at cap");
-        // recovery_key_sha256 trims; with no trailing whitespace the
-        // trimmed and untrimmed forms are equal.
-        let content_str = std::str::from_utf8(&content_bytes).unwrap();
-        let expected_digest = recovery_key_sha256(content_str);
-        replace_owner_only_secret_file(&src, &dst, &expected_digest)
-            .await
-            .expect("at-cap rename must succeed");
-        assert!(dst.exists());
-        assert!(!src.exists());
-        assert_eq!(std::fs::read(&dst).unwrap(), content_bytes);
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_started_recovery_key_rotation_refuses_unbound_pending_key() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let old_key = "recovery-key-before-started-marker";
-        let new_key = "recovery-key-pending-after-started-marker";
-        std::fs::write(&key_path, format!("{old_key}\n")).expect("write old key");
-        std::fs::write(&pending_path, format!("{new_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::Started,
-            None,
-            Some(recovery_key_sha256(old_key)),
-        )
-        .await
-        .expect("write started marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("started marker must not promote a pending key without a new-key digest");
-
-        assert!(
-            err.to_string()
-                .contains("started-stage marker never recorded a new pending key digest"),
-            "unexpected recovery error: {err}"
-        );
-        assert!(marker_path.exists());
-        assert!(pending_path.exists());
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{old_key}\n")
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_started_recovery_key_rotation_refuses_pending_when_current_missing() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(pending_path.parent().unwrap()).expect("create matrix dir");
-        let old_key = "recovery-key-before-started-marker";
-        let pending_key = "pending-key-after-started-marker";
-        std::fs::write(&pending_path, format!("{pending_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::Started,
-            None,
-            Some(recovery_key_sha256(old_key)),
-        )
-        .await
-        .expect("write started marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("started marker with missing current key must fail closed");
-
-        assert!(
-            err.to_string().contains("current key is missing"),
-            "unexpected recovery error: {err}"
-        );
-        assert!(pending_path.exists() && marker_path.exists());
-        assert!(!matrix_recovery_key_path(temp.path()).exists());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_recover_interrupted_recovery_key_rotation_refuses_stale_pending_over_restored_key(
-    ) {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let restored_key = "operator-restored-current-recovery-key";
-        let stale_pending_key = "stale-pending-recovery-key";
-        std::fs::write(&key_path, format!("{restored_key}\n")).expect("write restored key");
-        std::fs::write(&pending_path, format!("{stale_pending_key}\n"))
-            .expect("write stale pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::PendingKeyWritten,
-            Some(recovery_key_sha256(stale_pending_key)),
-            Some(recovery_key_sha256("previous-rotation-key")),
-        )
-        .await
-        .expect("write stale marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("stale pending key must not overwrite restored current key");
-
-        assert!(
-            err.to_string()
-                .contains("current key is neither the pre-rotation key nor the new pending key"),
-            "unexpected recovery error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{restored_key}\n"),
-            "current restored key must remain untouched"
-        );
-        assert!(
-            pending_path.exists() && marker_path.exists(),
-            "operator must inspect stale recovery artifacts explicitly"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_final_key_replaced_refuses_pending_over_restored_old_current_key() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let old_key = "operator-restored-old-recovery-key";
-        let new_key = "new-recovery-key-recorded-by-marker";
-        std::fs::write(&key_path, format!("{old_key}\n")).expect("write restored old key");
-        std::fs::write(&pending_path, format!("{new_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::FinalKeyReplaced,
-            Some(recovery_key_sha256(new_key)),
-            Some(recovery_key_sha256(old_key)),
-        )
-        .await
-        .expect("write final marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("final-stage marker must never promote pending over restored current key");
-
-        assert!(
-            err.to_string()
-                .contains("final-stage marker recorded key replacement"),
-            "unexpected recovery error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{old_key}\n")
-        );
-        assert!(pending_path.exists() && marker_path.exists());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_final_key_replaced_clears_stale_pending_when_current_matches_new_key() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let old_key = "recovery-key-before-rotation";
-        let new_key = "recovery-key-after-rotation";
-        std::fs::write(&key_path, format!("{new_key}\n")).expect("write final key");
-        std::fs::write(&pending_path, "stale-pending-material\n").expect("write stale pending");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::FinalKeyReplaced,
-            Some(recovery_key_sha256(new_key)),
-            Some(recovery_key_sha256(old_key)),
-        )
-        .await
-        .expect("write final marker");
-
-        recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect("final-stage marker with current new key should clear stale artifacts");
-
-        assert!(!pending_path.exists());
-        assert!(!marker_path.exists());
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{new_key}\n")
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_started_recovery_key_rotation_refuses_restored_current_key() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let previous_key = "previous-recovery-key";
-        let restored_key = "operator-restored-recovery-key";
-        let pending_key = "pending-recovery-key";
-        std::fs::write(&key_path, format!("{restored_key}\n")).expect("write restored key");
-        std::fs::write(&pending_path, format!("{pending_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::Started,
-            None,
-            Some(recovery_key_sha256(previous_key)),
-        )
-        .await
-        .expect("write started marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("started marker must not overwrite restored current key");
-
-        assert!(
-            err.to_string()
-                .contains("started-stage marker never recorded a new pending key digest"),
-            "unexpected recovery error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{restored_key}\n")
-        );
-        assert!(pending_path.exists() && marker_path.exists());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_legacy_recovery_key_rotation_marker_refuses_blind_promotion() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        std::fs::write(&key_path, "operator-restored\n").expect("write current key");
-        std::fs::write(&pending_path, "legacy-pending\n").expect("write pending key");
-        std::fs::write(&marker_path, "recovery-rotation-in-progress\n")
-            .expect("write legacy marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("legacy digest-less marker must fail closed");
-
-        assert!(
-            err.to_string()
-                .contains("does not record the previous local key digest"),
-            "unexpected recovery error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            "operator-restored\n"
-        );
-        assert!(pending_path.exists() && marker_path.exists());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_legacy_recovery_key_rotation_marker_without_current_key_still_fails_closed() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(pending_path.parent().unwrap()).expect("create matrix dir");
-        std::fs::write(&pending_path, "legacy-pending\n").expect("write pending key");
-        std::fs::write(&marker_path, "recovery-rotation-in-progress\n")
-            .expect("write legacy marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("legacy digest-less marker cannot prove pending ownership");
-
-        assert!(
-            err.to_string()
-                .contains("does not record the previous local key digest"),
-            "unexpected recovery error: {err}"
-        );
-        assert!(pending_path.exists() && marker_path.exists());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_pending_key_written_refuses_restored_current_mismatch() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let key_path = matrix_recovery_key_path(temp.path());
-        let pending_path = matrix_recovery_pending_key_path(temp.path());
-        let marker_path = matrix_recovery_rotating_marker_path(temp.path());
-        std::fs::create_dir_all(key_path.parent().unwrap()).expect("create matrix dir");
-        let previous_key = "previous-recovery-key";
-        let restored_key = "operator-restored-current-key";
-        let pending_key = "new-pending-key";
-        std::fs::write(&key_path, format!("{restored_key}\n")).expect("write restored key");
-        std::fs::write(&pending_path, format!("{pending_key}\n")).expect("write pending key");
-        write_recovery_rotation_marker_stage_durable(
-            &marker_path,
-            RecoveryKeyRotationMarkerStage::PendingKeyWritten,
-            Some(recovery_key_sha256(pending_key)),
-            Some(recovery_key_sha256(previous_key)),
-        )
-        .await
-        .expect("write pending marker");
-
-        let err = recover_interrupted_recovery_key_rotation(temp.path())
-            .await
-            .expect_err("pending key must not replace restored current mismatch");
-
-        assert!(
-            err.to_string()
-                .contains("current key is neither the pre-rotation key nor the new pending key"),
-            "unexpected recovery error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&key_path).unwrap(),
-            format!("{restored_key}\n")
-        );
-        assert!(pending_path.exists() && marker_path.exists());
     }
 
     /// Drive 3 consecutive bad-event_id failures through the
@@ -14710,431 +11473,6 @@ mod tests {
             Some("session-history-corrupt"),
             "pending inbound failures must retain their typed routing kind"
         );
-    }
-
-    /// Encryption-flag flip detection: a plaintext record (encoded
-    /// when `matrix.encrypted=false`) MUST decode under a config
-    /// that has since flipped to `matrix.encrypted=true`. The
-    /// reverse direction (encrypted line under encrypted=false
-    /// config) requires the original passphrase to decrypt, so it
-    /// is ALSO a supported migration but only when the operator
-    /// keeps the passphrase configured during the transition —
-    /// the introspection branch correctly recognizes the line
-    /// shape, but key resolution still has to succeed. This test
-    /// pins the introspection logic itself: line shape governs
-    /// the decode branch, NOT `config.encrypted()`.
-    #[test]
-    fn test_matrix_inbound_dlq_plaintext_decodes_under_encrypted_config() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let record = matrix_test_dlq_record();
-        let plain_config = matrix_test_config(false);
-        let enc_config = matrix_test_config(true);
-
-        // Encode under encrypted=false, decode under encrypted=true.
-        // The plaintext line carries no version/nonce/ciphertext,
-        // so the introspection branch falls into plaintext-decode
-        // regardless of the config's encryption flag.
-        let plain_line = encode_matrix_inbound_dlq_record(temp.path(), &plain_config, &record)
-            .expect("plaintext encode");
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &enc_config, &plain_line)
-            .expect("plaintext line must decode under encrypted config via introspection");
-        assert_eq!(decoded, record);
-    }
-
-    #[test]
-    fn test_temporarily_undecodable_encrypted_dlq_error_is_preserved() {
-        assert!(is_temporarily_undecodable_dlq_error(
-            &MatrixError::MissingStoreSecret
-        ));
-        assert!(is_temporarily_undecodable_dlq_error(
-            &MatrixError::SyncFailed(
-                "encrypted v2 DLQ record encountered but no key cache or config available"
-                    .to_string(),
-            )
-        ));
-        assert!(!is_temporarily_undecodable_dlq_error(
-            &MatrixError::SyncFailed("Matrix inbound DLQ corrupt record".to_string())
-        ));
-        // Post-Batch-79: `LegacyDlqEnvelopeRefused` is the operator's
-        // EXPLICIT policy choice and no toggle makes the records
-        // decodable later, so it is NOT a temporarily-undecodable
-        // class. Routing it through the Corrupt branch preserves
-        // refused records in the quarantine artifact for operator
-        // inspection rather than the live-DLQ tail where cap-pressure
-        // would drop them.
-        assert!(!is_temporarily_undecodable_dlq_error(
-            &MatrixError::LegacyDlqEnvelopeRefused
-        ));
-    }
-
-    #[test]
-    fn test_rotate_matrix_inbound_dlq_for_rekey_keeps_old_backup_until_cleanup() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let old_config = matrix_test_config_with_passphrase(&old_passphrase);
-        let new_config = matrix_test_config_with_passphrase(&new_passphrase);
-        let record = matrix_test_dlq_record();
-        let old_line = encode_matrix_inbound_dlq_record(temp.path(), &old_config, &record)
-            .expect("old-keyed DLQ line");
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&live_path, format!("{old_line}\n")).expect("write live DLQ");
-
-        let outcome = rotate_matrix_inbound_dlq_for_rekey(
-            temp.path(),
-            &old_config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect("rekey live DLQ");
-
-        let MatrixDlqRekeyOutcome::Rotated {
-            decoded_count,
-            backup_path,
-        } = outcome
-        else {
-            panic!("non-empty DLQ must rotate");
-        };
-        assert_eq!(decoded_count, 1);
-        assert!(
-            backup_path.exists(),
-            "old-keyed backup must remain until passphrase promotion succeeds"
-        );
-        let new_live = std::fs::read_to_string(&live_path).expect("read new live DLQ");
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &new_config, new_live.trim())
-            .expect("new live DLQ must decode under new passphrase");
-        assert_eq!(decoded, record);
-        let old_backup = std::fs::read_to_string(&backup_path).expect("read backup DLQ");
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &old_config, old_backup.trim())
-            .expect("backup DLQ must remain under old passphrase");
-        assert_eq!(decoded, record);
-    }
-
-    #[test]
-    fn test_rotate_matrix_inbound_dlq_for_rekey_honors_legacy_refuse_policy() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let mut old_config = matrix_test_config_with_passphrase(&old_passphrase);
-        old_config.legacy_dlq_envelope_policy = MatrixLegacyDlqEnvelopePolicy::Refuse;
-        let record = matrix_test_dlq_record();
-        let v1_line =
-            encode_legacy_v1_matrix_inbound_dlq_record_for_test(temp.path(), &old_config, &record);
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&live_path, format!("{v1_line}\n")).expect("write live DLQ");
-
-        let err = rotate_matrix_inbound_dlq_for_rekey(
-            temp.path(),
-            &old_config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect_err("rekey must not launder refused v1 into v2");
-
-        assert!(matches!(err, MatrixError::LegacyDlqEnvelopeRefused));
-        assert!(
-            err.to_string()
-                .contains("legacy Matrix inbound DLQ v1 envelope refused by policy"),
-            "unexpected rekey refusal error: {err}"
-        );
-        assert!(
-            matrix_inbound_dlq_rekey_backup_path(temp.path())
-                .try_exists()
-                .is_ok_and(|exists| !exists),
-            "refused rekey must leave the original live file in place"
-        );
-    }
-
-    #[test]
-    fn test_rotate_matrix_inbound_dlq_for_rekey_enforces_cap_before_decode() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let config = matrix_test_config_with_passphrase(&old_passphrase);
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        let mut content = String::new();
-        for _ in 0..=MATRIX_INBOUND_DLQ_MAX_RECORDS {
-            content.push_str("{}\n");
-        }
-        std::fs::write(&live_path, content).expect("write over-cap live DLQ");
-
-        let err = rotate_matrix_inbound_dlq_for_rekey(
-            temp.path(),
-            &config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect_err("over-cap DLQ must fail before decrypting malformed records");
-
-        assert!(
-            err.to_string().contains("inbound DLQ has more than"),
-            "unexpected error: {err}"
-        );
-    }
-
-    /// A planted regular DLQ file passing the no-follow open check
-    /// but holding one huge newline-free line used to OOM the rekey
-    /// reader (unbounded `BufRead::read_line` into a String). The
-    /// per-line cap mirrors the live-DLQ replay/count seam at
-    /// `read_matrix_inbound_dlq_lines_streaming` and
-    /// `matrix_inbound_dlq_line_count` — the rekey reader must fail
-    /// closed at the same threshold, not allocate the whole line.
-    #[test]
-    fn test_rotate_matrix_inbound_dlq_for_rekey_enforces_per_line_cap() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let config = matrix_test_config_with_passphrase(&old_passphrase);
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        // One huge newline-free line just over the per-line cap. No
-        // terminating newline — the cap-without-newline branch fails
-        // closed before any further allocation.
-        let oversize = "x".repeat(MATRIX_INBOUND_DLQ_REPLAY_LINE_MAX_BYTES + 1);
-        std::fs::write(&live_path, &oversize).expect("write oversize line DLQ");
-
-        let err = rotate_matrix_inbound_dlq_for_rekey(
-            temp.path(),
-            &config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect_err("oversize-line DLQ must fail before allocating the full line");
-
-        assert!(
-            err.to_string().contains(&format!(
-                "exceeding {MATRIX_INBOUND_DLQ_REPLAY_LINE_MAX_BYTES} bytes"
-            )),
-            "unexpected error (expected per-line cap message): {err}"
-        );
-        assert!(
-            matrix_inbound_dlq_rekey_backup_path(temp.path())
-                .try_exists()
-                .is_ok_and(|exists| !exists),
-            "refused rekey must leave the original live file in place"
-        );
-    }
-
-    #[test]
-    fn test_recover_matrix_inbound_dlq_rekey_restores_from_backup_when_live_empty() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let old_config = matrix_test_config_with_passphrase(&old_passphrase);
-        let new_config = matrix_test_config_with_passphrase(&new_passphrase);
-        let record = matrix_test_dlq_record();
-        let old_line = encode_matrix_inbound_dlq_record(temp.path(), &old_config, &record)
-            .expect("old-keyed DLQ line");
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        let backup_path = matrix_inbound_dlq_rekey_backup_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&live_path, "").expect("write empty live DLQ");
-        std::fs::write(&backup_path, format!("{old_line}\n")).expect("write backup DLQ");
-
-        let outcome = recover_matrix_inbound_dlq_rekey(
-            temp.path(),
-            &new_config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect("recover DLQ rekey");
-
-        let MatrixDlqRekeyOutcome::Rotated {
-            decoded_count,
-            backup_path: outcome_backup,
-        } = outcome
-        else {
-            panic!("backup-only DLQ recovery must rotate");
-        };
-        assert_eq!(decoded_count, 1);
-        assert_eq!(outcome_backup, backup_path);
-        assert!(
-            backup_path.exists(),
-            "backup must remain until the pending passphrase is promoted"
-        );
-        let new_live = std::fs::read_to_string(&live_path).expect("read recovered live DLQ");
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &new_config, new_live.trim())
-            .expect("recovered live DLQ must decode under new passphrase");
-        assert_eq!(decoded, record);
-    }
-
-    #[test]
-    fn test_recover_matrix_inbound_dlq_rekey_honors_legacy_refuse_policy() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let old_config = matrix_test_config_with_passphrase(&old_passphrase);
-        let mut new_config = matrix_test_config_with_passphrase(&new_passphrase);
-        new_config.legacy_dlq_envelope_policy = MatrixLegacyDlqEnvelopePolicy::Refuse;
-        let record = matrix_test_dlq_record();
-        let v1_line =
-            encode_legacy_v1_matrix_inbound_dlq_record_for_test(temp.path(), &old_config, &record);
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        let backup_path = matrix_inbound_dlq_rekey_backup_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&live_path, "").expect("write empty live DLQ");
-        std::fs::write(&backup_path, format!("{v1_line}\n")).expect("write backup DLQ");
-
-        let err = recover_matrix_inbound_dlq_rekey(
-            temp.path(),
-            &new_config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect_err("rekey recovery must not launder refused v1 into v2");
-
-        assert!(matches!(err, MatrixError::LegacyDlqEnvelopeRefused));
-        assert!(
-            err.to_string()
-                .contains("legacy Matrix inbound DLQ v1 envelope refused by policy"),
-            "unexpected rekey recovery refusal error: {err}"
-        );
-        assert!(
-            backup_path.exists(),
-            "refused recovery must keep the old backup for operator policy reversal"
-        );
-    }
-
-    #[test]
-    fn test_recover_matrix_inbound_dlq_rekey_keeps_rekeyed_live_when_backup_exists() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let old_config = matrix_test_config_with_passphrase(&old_passphrase);
-        let new_config = matrix_test_config_with_passphrase(&new_passphrase);
-        let record = matrix_test_dlq_record();
-        let old_line = encode_matrix_inbound_dlq_record(temp.path(), &old_config, &record)
-            .expect("old-keyed DLQ line");
-        let new_line = encode_matrix_inbound_dlq_record(temp.path(), &new_config, &record)
-            .expect("new-keyed DLQ line");
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        let backup_path = matrix_inbound_dlq_rekey_backup_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&live_path, format!("{new_line}\n")).expect("write rekeyed live DLQ");
-        std::fs::write(&backup_path, format!("{old_line}\n")).expect("write old backup DLQ");
-
-        let outcome = recover_matrix_inbound_dlq_rekey(
-            temp.path(),
-            &new_config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect("recover DLQ rekey");
-
-        let MatrixDlqRekeyOutcome::Rotated {
-            decoded_count,
-            backup_path: outcome_backup,
-        } = outcome
-        else {
-            panic!("backup+live recovery should keep rekeyed live as rotated");
-        };
-        assert_eq!(decoded_count, 1);
-        assert_eq!(outcome_backup, backup_path);
-        assert!(
-            backup_path.exists(),
-            "old backup must remain until pending passphrase promotion succeeds"
-        );
-        let live_after = std::fs::read_to_string(&live_path).expect("read live DLQ");
-        let decoded = decode_matrix_inbound_dlq_record(temp.path(), &new_config, live_after.trim())
-            .expect("live DLQ must remain decodable under new passphrase");
-        assert_eq!(decoded, record);
-    }
-
-    #[test]
-    fn test_recover_matrix_inbound_dlq_rekey_refuses_old_keyed_live_with_backup() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let old_passphrase = crate::crypto::generate_hex_secret(32).expect("old passphrase");
-        let new_passphrase = crate::crypto::generate_hex_secret(32).expect("new passphrase");
-        let old_config = matrix_test_config_with_passphrase(&old_passphrase);
-        let record = matrix_test_dlq_record();
-        let old_line = encode_matrix_inbound_dlq_record(temp.path(), &old_config, &record)
-            .expect("old-keyed DLQ line");
-        let live_path = matrix_inbound_dlq_path(temp.path());
-        let backup_path = matrix_inbound_dlq_rekey_backup_path(temp.path());
-        std::fs::create_dir_all(live_path.parent().expect("DLQ parent")).expect("create parent");
-        std::fs::write(&live_path, format!("{old_line}\n")).expect("write old-keyed live DLQ");
-        std::fs::write(&backup_path, format!("{old_line}\n")).expect("write old backup DLQ");
-
-        let new_config = matrix_test_config_with_passphrase(&new_passphrase);
-        let err = recover_matrix_inbound_dlq_rekey(
-            temp.path(),
-            &new_config,
-            &old_passphrase,
-            &new_passphrase,
-        )
-        .expect_err("backup recovery must not clobber a non-empty live DLQ");
-
-        assert!(
-            err.to_string().contains("refusing to clobber"),
-            "unexpected rekey recovery error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&live_path).expect("read live after refusal"),
-            format!("{old_line}\n")
-        );
-    }
-
-    /// `to-device` verification events from neither the
-    /// configured user nor the auto-join allowlist must be dropped
-    /// without entering the verification record store. Otherwise a
-    /// hostile peer can spam 256 fresh transaction_ids and evict
-    /// the operator's legitimate flow at index 0 (the cap-eviction
-    /// fallback path).
-    ///
-    /// The handler is async and consumes an SDK event type that's
-    /// awkward to construct in unit tests, so this test exercises
-    /// the GATE policy directly via the helper
-    /// `MatrixAutoJoinConfig::allows_user` and the self-equality
-    /// check that the handler uses. A regression that loosens the
-    /// gate (e.g. removes the `allows_user` arm) would not trip
-    /// this test, but a regression that breaks the helpers
-    /// themselves would, and a static-analysis pin against the
-    /// handler body catches the gate-removal class of refactor.
-    #[test]
-    fn test_handle_to_device_verification_gate_helpers_reject_unallowed_peer() {
-        let config = matrix_test_config(false);
-        // Non-allowlisted peer, not the configured user.
-        assert!(
-            !config.auto_join.allows_user("@hostile-peer:evil.com"),
-            "default test config must not allowlist arbitrary peers"
-        );
-        // Configured user is `@cara:example.com` per matrix_test_config.
-        assert_eq!(config.user_id, "@cara:example.com");
-
-        // Pin the source body — the handler's gate must combine
-        // self-equality OR allowlist; a refactor that drops either
-        // arm breaks the contract.
-        let body = matrix_rs_fn_body("async fn handle_to_device_event");
-        let body = body.as_str();
-        assert!(
-            body.contains("matrix_user_ids_equal(sender, &config.user_id)")
-                || body.contains("matrix_user_ids_equal(\n"),
-            "handle_to_device_event must check self-verification via matrix_user_ids_equal"
-        );
-        assert!(
-            body.contains("auto_join.allows_user"),
-            "handle_to_device_event must consult the allowlist for non-self peers"
-        );
-        assert!(
-            body.contains("to-device verification event dropped"),
-            "handle_to_device_event must drop unallowed peers (warn-log marker present)"
-        );
-        let classifier = matrix_rs_fn_body("fn matrix_to_device_verification_sender_and_kind");
-        let classifier = classifier.as_str();
-        for event_type in [
-            "m.key.verification.start",
-            "m.key.verification.ready",
-            "m.key.verification.key",
-            "m.key.verification.mac",
-        ] {
-            assert!(
-                classifier.contains(event_type),
-                "to-device verification classifier must include {event_type}"
-            );
-        }
     }
 
     /// Static-analysis pin for `handle_invites`: a non-allowlisted
