+++ src/channels/matrix/verification.rs
@@ -1,3 +1,11 @@
+//! Matrix SAS verification state.
+//!
+//! This module owns daemon-side verification records, Matrix protocol flow
+//! lookup, and the control actions that advance SAS handshakes. Records keep
+//! sanitized identifiers for operator-visible JSON/log surfaces and raw
+//! identifiers for SDK lookups because Matrix transaction/user IDs are
+//! protocol keys, not display strings.
+
 use super::*;
 use matrix_sdk::encryption::verification::{SasState, SasVerification, VerificationRequestState};
 
@@ -21,7 +29,14 @@ pub struct MatrixVerificationInfo {
     /// bytes from the original to-device event).
     #[serde(skip)]
     pub raw_protocol_flow_id: String,
+    /// Matrix user id sanitized for operator-visible surfaces.
     pub user_id: String,
+    /// Raw Matrix user id as supplied by the SDK. The SDK indexes
+    /// verification state by the raw user id and raw protocol flow id;
+    /// sanitize is non-bijective and must not feed lookups or
+    /// same-peer equality.
+    #[serde(skip)]
+    pub raw_user_id: String,
     /// Device id of the peer being verified, or absent when the
     /// protocol flow targets the user without a specific device.
     /// `skip_serializing_if = Option::is_none` matches the convention
@@ -238,7 +253,7 @@ pub(super) fn matrix_to_device_verification_sender_and_kind(
 /// whose request already arrived inbound would re-broadcast
 /// `requested` on top of the inbound-handler's broadcast, duplicating
 /// UI notifications.
-pub(crate) struct MatrixStartVerificationOutcome {
+pub(super) struct MatrixStartVerificationOutcome {
     pub info: MatrixVerificationInfo,
     pub inserted: bool,
 }
@@ -246,7 +261,7 @@ pub(crate) struct MatrixStartVerificationOutcome {
 #[derive(Debug, Clone)]
 pub(super) enum VerificationRecordUpsert {
     Applied {
-        info: MatrixVerificationInfo,
+        info: Box<MatrixVerificationInfo>,
         inserted: bool,
     },
     RejectedAtCap,
@@ -256,7 +271,7 @@ impl VerificationRecordUpsert {
     #[cfg(test)]
     pub(super) fn unwrap_applied(self) -> (MatrixVerificationInfo, bool) {
         match self {
-            Self::Applied { info, inserted } => (info, inserted),
+            Self::Applied { info, inserted } => (*info, inserted),
             Self::RejectedAtCap => panic!("verification record upsert unexpectedly hit the cap"),
         }
     }
@@ -353,7 +368,10 @@ pub(super) async fn start_matrix_verification(
                 .to_string(),
         ));
     };
-    Ok(MatrixStartVerificationOutcome { info, inserted })
+    Ok(MatrixStartVerificationOutcome {
+        info: *info,
+        inserted,
+    })
 }
 
 pub(super) fn matrix_verification_control_id(user_id: &str, protocol_flow_id: &str) -> String {
@@ -399,6 +417,7 @@ pub(super) fn upsert_verification_record(
         flow.protocol_flow_id = protocol_flow_id;
         flow.raw_protocol_flow_id = raw_protocol_flow_id;
         flow.user_id = user_id;
+        flow.raw_user_id = raw_user_id;
         flow.device_id = device_id;
         flow.state = flow_state;
         // Preserve any previously captured SAS data on re-upsert. A
@@ -411,10 +430,11 @@ pub(super) fn upsert_verification_record(
         flow.updated_at = now;
         let flow = flow.clone();
         return VerificationRecordUpsert::Applied {
-            info: flow,
+            info: Box::new(flow),
             inserted: false,
         };
     }
+
     // Enforce a hard cap before insert so a flood of fresh flow_ids
     // (allowlisted peer spam, redelivery storm) cannot grow the Vec
     // unbounded between TTL prunes. Eviction priority: terminal records
@@ -428,7 +448,7 @@ pub(super) fn upsert_verification_record(
             .position(|f| f.state.is_terminal())
             .or_else(|| {
                 guard.verifications.iter().position(|f| {
-                    f.state == MatrixVerificationState::Requested && f.user_id == user_id
+                    f.state == MatrixVerificationState::Requested && f.raw_user_id == raw_user_id
                 })
             })
         else {
@@ -466,6 +486,7 @@ pub(super) fn upsert_verification_record(
         protocol_flow_id,
         raw_protocol_flow_id,
         user_id,
+        raw_user_id,
         device_id,
         state: flow_state,
         sas: None,
@@ -474,7 +495,7 @@ pub(super) fn upsert_verification_record(
     };
     guard.verifications.push(flow.clone());
     VerificationRecordUpsert::Applied {
-        info: flow,
+        info: Box::new(flow),
         inserted: true,
     }
 }
@@ -507,7 +528,7 @@ pub(super) async fn apply_verification_action(
     let audit_successful_confirm =
         matches!(action, MatrixVerificationAction::Confirm { matches: true });
     let parsed_user_id: OwnedUserId = flow
-        .user_id
+        .raw_user_id
         .parse::<OwnedUserId>()
         .map_err(|err| MatrixError::InvalidUserId(err.to_string()))?;
     // SDK lookups must use the RAW flow id from the original
@@ -790,7 +811,7 @@ pub(super) async fn refresh_verification_records(
     let mut invalid_user_id_warn_count = 0usize;
     let mut suppressed_invalid_user_id_warn_count = 0usize;
     for record in records {
-        let parsed_user_id: OwnedUserId = match record.user_id.parse() {
+        let parsed_user_id: OwnedUserId = match record.raw_user_id.parse() {
             Ok(user_id) => user_id,
             Err(err) => {
                 // Skip the record rather than aborting the loop. With
@@ -959,4 +980,957 @@ pub(super) fn prune_finished_verification_records(state: &Arc<RwLock<MatrixRunti
 pub(super) fn matrix_verification_cap_warn_should_fire() -> bool {
     static LAST_WARN_AT_SECS: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
     crate::logging::throttle::throttled_once_per_hour(&LAST_WARN_AT_SECS)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::channels::matrix::matrix_rs_fn_body;
+    use serde_json::json;
+
+    /// `update_verification_record_state` must preserve previously
+    /// captured SAS data when the caller passes `SasUpdate::Preserve`
+    /// — regression test pinning that emoji stay visible across state
+    /// transitions instead of being clobbered on every refresh
+    /// iteration.
+    #[test]
+    fn test_update_verification_record_state_preserves_sas_when_next_is_none() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (info, inserted) = upsert_verification_record(
+            &state,
+            "abc123".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        let flow_id = info.flow_id.clone();
+        let seeded_sas = MatrixSasInfo {
+            emoji: Some(vec![MatrixSasEmoji {
+                symbol: "🐱".to_string(),
+                description: "Cat".to_string(),
+            }]),
+            decimals: Some([1, 2, 3]),
+        };
+        update_verification_record_state(
+            &state,
+            &flow_id,
+            MatrixVerificationState::KeysExchanged,
+            SasUpdate::Set(seeded_sas.clone()),
+        )
+        .expect("seed SAS");
+
+        // Preserve must keep the SAS intact even though the state advances.
+        update_verification_record_state(
+            &state,
+            &flow_id,
+            MatrixVerificationState::Confirmed,
+            SasUpdate::Preserve,
+        )
+        .expect("preserve");
+        let stored = state
+            .read()
+            .verifications
+            .iter()
+            .find(|f| f.flow_id == flow_id)
+            .cloned()
+            .expect("flow exists");
+        assert_eq!(stored.state, MatrixVerificationState::Confirmed);
+        assert_eq!(stored.sas, Some(seeded_sas.clone()));
+
+        // Set replaces.
+        let replacement = MatrixSasInfo {
+            emoji: None,
+            decimals: Some([9, 8, 7]),
+        };
+        update_verification_record_state(
+            &state,
+            &flow_id,
+            MatrixVerificationState::Done,
+            SasUpdate::Set(replacement.clone()),
+        )
+        .expect("set");
+        let stored = state
+            .read()
+            .verifications
+            .iter()
+            .find(|f| f.flow_id == flow_id)
+            .cloned()
+            .expect("flow exists");
+        assert_eq!(stored.sas, Some(replacement));
+    }
+
+    /// `update_verification_record_state` returns `Ok(None)` when the
+    /// requested next state and SAS are equal to the existing record
+    /// (no-op tick) — refresh broadcasts only fire on real change;
+    /// this test pins that contract.
+    #[test]
+    fn test_update_verification_record_state_returns_none_on_noop() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (info, inserted) = upsert_verification_record(
+            &state,
+            "noop-flow".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        let flow_id = info.flow_id.clone();
+        // Same state, SasUpdate::Preserve (so SAS stays None) — must
+        // return Ok(None) so the broadcast caller knows nothing changed.
+        let result = update_verification_record_state(
+            &state,
+            &flow_id,
+            MatrixVerificationState::Requested,
+            SasUpdate::Preserve,
+        )
+        .expect("call");
+        assert!(
+            result.is_none(),
+            "no-op state update must return None so refresh broadcasts only fire on real change"
+        );
+    }
+
+    /// A real state change must return `Ok(Some(record))` so the
+    /// caller can broadcast `matrix.verification.updated` with the
+    /// post-state record.
+    #[test]
+    fn test_update_verification_record_state_returns_some_on_change() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (info, _) = upsert_verification_record(
+            &state,
+            "change-flow".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        let flow_id = info.flow_id.clone();
+        let result = update_verification_record_state(
+            &state,
+            &flow_id,
+            MatrixVerificationState::KeysExchanged,
+            SasUpdate::Preserve,
+        )
+        .expect("call");
+        let updated = result.expect("real change must return Some");
+        assert_eq!(updated.state, MatrixVerificationState::KeysExchanged);
+    }
+
+    /// A SAS-only change with the same state must still report the
+    /// change. Otherwise a refresh that captures fresh SAS data on a
+    /// flow that's been in `KeysExchanged` for a while would fail to
+    /// broadcast and the operator UI would never see the comparison
+    /// emoji.
+    #[test]
+    fn test_update_verification_record_state_returns_some_when_sas_changes() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (info, _) = upsert_verification_record(
+            &state,
+            "sas-flow".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE".to_string()),
+            MatrixVerificationState::KeysExchanged,
+        )
+        .unwrap_applied();
+        let flow_id = info.flow_id.clone();
+        let new_sas = MatrixSasInfo {
+            emoji: None,
+            decimals: Some([1, 2, 3]),
+        };
+        let result = update_verification_record_state(
+            &state,
+            &flow_id,
+            MatrixVerificationState::KeysExchanged,
+            SasUpdate::Set(new_sas.clone()),
+        )
+        .expect("call");
+        let updated = result.expect("SAS change must return Some");
+        assert_eq!(updated.sas, Some(new_sas));
+    }
+
+    /// Pin the JSON shape of `MatrixVerificationInfo` against the
+    /// schema declared in `tests/golden/ws/events.json`. If the
+    /// runtime adds/removes a public field on the verification info,
+    /// the golden schema must be updated in lockstep — this test
+    /// catches drift before WS clients see it.
+    #[test]
+    fn test_matrix_verification_info_serializes_required_fields() {
+        let info = MatrixVerificationInfo {
+            flow_id: "flow-1".to_string(),
+            protocol_flow_id: "txn-1".to_string(),
+            raw_protocol_flow_id: "txn-1".to_string(),
+            user_id: "@alice:example.com".to_string(),
+            raw_user_id: "@alice:example.com".to_string(),
+            device_id: Some("DEVICE".to_string()),
+            state: MatrixVerificationState::Requested,
+            sas: None,
+            created_at: 1,
+            updated_at: 2,
+        };
+        let json = serde_json::to_value(&info).expect("serialize");
+        // Required fields per the WS event schema.
+        for field in [
+            "flowId",
+            "protocolFlowId",
+            "userId",
+            "state",
+            "createdAt",
+            "updatedAt",
+        ] {
+            assert!(
+                json.get(field).is_some(),
+                "MatrixVerificationInfo must serialize {field}"
+            );
+        }
+        // State must serialize to one of the documented values.
+        let state_str = json
+            .get("state")
+            .and_then(|v| v.as_str())
+            .expect("state is string");
+        let allowed = [
+            "created",
+            "requested",
+            "ready",
+            "transitioned",
+            "started",
+            "accepted",
+            "keys_exchanged",
+            "confirmed",
+            "done",
+            "cancelled",
+            "mismatched",
+        ];
+        assert!(
+            allowed.contains(&state_str),
+            "MatrixVerificationState '{state_str}' is not in the documented enum"
+        );
+    }
+
+    /// Pin: `upsert_verification_record` stores BOTH sanitized
+    /// identifiers (operator-display surface) and raw identifiers
+    /// (`raw_*`, SDK lookup keys / internal equality). Sanitization is
+    /// non-bijective; passing the sanitized form to
+    /// `client.encryption().get_verification_*` would fail to
+    /// resolve any flow whose original id contained codepoints
+    /// stripped by sanitize. `apply_verification_action` MUST use
+    /// the raw form for SDK lookups.
+    #[test]
+    fn test_upsert_verification_record_preserves_raw_protocol_fields() {
+        let runtime_state = Arc::new(parking_lot::RwLock::new(MatrixRuntimeState::default()));
+        // Inject a hostile-shape protocol flow id with a zero-width
+        // joiner. Sanitize strips ZWJs (U+200D); the raw must be
+        // preserved exactly so SDK lookup matches.
+        let raw_flow = "txn-\u{200d}-abc";
+        let sanitized_flow = "txn--abc";
+        let raw_user_id = "@ali\u{200d}ce:example.com";
+        let sanitized_user_id = "@alice:example.com";
+        let (info, _inserted) = upsert_verification_record(
+            &runtime_state,
+            raw_flow.to_string(),
+            raw_user_id.to_string(),
+            Some("DEVICE".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert_eq!(
+            info.raw_protocol_flow_id, raw_flow,
+            "raw_protocol_flow_id must preserve the original bytes for SDK lookup"
+        );
+        assert_eq!(
+            info.protocol_flow_id, sanitized_flow,
+            "protocol_flow_id must be sanitized for operator-display surfaces"
+        );
+        assert_eq!(
+            info.raw_user_id, raw_user_id,
+            "raw_user_id must preserve the original bytes for SDK lookup and equality"
+        );
+        assert_eq!(
+            info.user_id, sanitized_user_id,
+            "user_id must be sanitized for operator-display surfaces"
+        );
+        // Wire serialization MUST omit raw_* fields (they
+        // would defeat the whole point: operator scripts decoding the
+        // wire JSON would see un-sanitized bytes).
+        let json = serde_json::to_value(&info).expect("serialize");
+        assert!(
+            json.get("rawProtocolFlowId").is_none() && json.get("raw_protocol_flow_id").is_none(),
+            "raw_protocol_flow_id must NOT serialize to wire JSON"
+        );
+        assert!(
+            json.get("rawUserId").is_none() && json.get("raw_user_id").is_none(),
+            "raw_user_id must NOT serialize to wire JSON"
+        );
+    }
+
+    /// Pin: `apply_verification_action` resolves SDK lookups via the
+    /// `raw_protocol_flow_id` field, never the sanitized form.
+    /// Static-analysis pin against the function body: the let-binding
+    /// must be `flow.raw_protocol_flow_id.clone()` and `protocol_flow_id`
+    /// must NOT be re-bound to `flow.protocol_flow_id` (the sanitized
+    /// form) anywhere in the SDK-call section.
+    #[test]
+    fn test_apply_verification_action_uses_raw_for_sdk_lookup() {
+        let body = matrix_rs_fn_body("async fn apply_verification_action");
+        let body = body.as_str();
+        assert!(
+            body.contains("let protocol_flow_id = flow.raw_protocol_flow_id.clone();"),
+            "apply_verification_action must clone raw_protocol_flow_id for SDK lookup"
+        );
+        assert!(
+            body.contains(".raw_user_id\n        .parse::<OwnedUserId>()"),
+            "apply_verification_action must parse raw_user_id for SDK lookup"
+        );
+        assert!(
+            !body.contains("let protocol_flow_id = flow.protocol_flow_id.clone();"),
+            "apply_verification_action must NOT re-bind protocol_flow_id to the \
+             sanitized field; sanitize is non-bijective and SDK lookup would fail"
+        );
+        assert!(
+            !body.contains(".user_id\n        .parse::<OwnedUserId>()"),
+            "apply_verification_action must not parse sanitized user_id for SDK lookup"
+        );
+    }
+
+    /// Pin: `refresh_verification_records` also resolves SDK
+    /// lookups via the raw protocol flow id. A ZWSP / bidi /
+    /// control-stripped flow can sit in the daemon list long enough
+    /// for the refresh worker to update it; using the sanitized
+    /// display field there would make refresh mark the flow stale
+    /// while `apply_verification_action` still finds it.
+    #[test]
+    fn test_refresh_verification_records_uses_raw_for_sdk_lookup() {
+        let body = matrix_rs_fn_body("async fn refresh_verification_records");
+        let body = body.as_str();
+        assert!(
+            body.contains(
+                "get_verification_request(&parsed_user_id, &record.raw_protocol_flow_id)"
+            ),
+            "refresh_verification_records must use raw_protocol_flow_id for request lookup"
+        );
+        assert!(
+            body.contains("get_verification(&parsed_user_id, &record.raw_protocol_flow_id)"),
+            "refresh_verification_records must use raw_protocol_flow_id for SAS lookup"
+        );
+        assert!(
+            body.contains("record.raw_user_id.parse()"),
+            "refresh_verification_records must parse raw_user_id for SDK lookup"
+        );
+        assert!(
+            !body.contains("get_verification_request(&parsed_user_id, &record.protocol_flow_id)")
+                && !body.contains("get_verification(&parsed_user_id, &record.protocol_flow_id)"),
+            "refresh_verification_records must not use the sanitized protocol_flow_id for SDK lookups"
+        );
+        assert!(
+            !body.contains("record.user_id.parse()"),
+            "refresh_verification_records must not parse sanitized user_id for SDK lookup"
+        );
+    }
+
+    #[test]
+    fn test_verification_record_upsert_and_prune() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (first, inserted) = upsert_verification_record(
+            &state,
+            "protocol-flow-1".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE1".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        assert_eq!(first.protocol_flow_id, "protocol-flow-1");
+        assert_eq!(first.state, MatrixVerificationState::Requested);
+
+        let (updated, inserted) = upsert_verification_record(
+            &state,
+            "protocol-flow-1".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE2".to_string()),
+            MatrixVerificationState::Ready,
+        )
+        .unwrap_applied();
+        assert!(!inserted);
+        assert_eq!(updated.device_id.as_deref(), Some("DEVICE2"));
+        assert_eq!(updated.state, MatrixVerificationState::Ready);
+
+        state.write().verifications[0].updated_at =
+            now_millis() - MATRIX_VERIFICATION_RECORD_TTL.as_millis() as i64 - 1;
+        prune_verification_records(&state);
+        assert!(state.read().verifications.is_empty());
+    }
+
+    #[test]
+    fn test_matrix_verification_info_wire_shape_is_camel_case_and_raw_safe() {
+        let info = MatrixVerificationInfo {
+            flow_id: "mvr_test".to_string(),
+            protocol_flow_id: "txn-safe".to_string(),
+            raw_protocol_flow_id: "txn-\u{200b}-raw".to_string(),
+            user_id: "@alice:example.com".to_string(),
+            raw_user_id: "@alice:example.com".to_string(),
+            device_id: Some("DEVICE".to_string()),
+            state: MatrixVerificationState::KeysExchanged,
+            sas: Some(MatrixSasInfo {
+                emoji: Some(vec![MatrixSasEmoji {
+                    symbol: "🐱".to_string(),
+                    description: "Cat".to_string(),
+                }]),
+                decimals: Some([1234, 5678, 9012]),
+            }),
+            created_at: 10,
+            updated_at: 20,
+        };
+
+        let json = serde_json::to_value(&info).expect("serialize verification info");
+        assert_eq!(json["flowId"], "mvr_test");
+        assert_eq!(json["protocolFlowId"], "txn-safe");
+        assert!(json.get("rawProtocolFlowId").is_none());
+        assert!(json.get("raw_protocol_flow_id").is_none());
+        assert_eq!(json["userId"], "@alice:example.com");
+        assert!(json.get("rawUserId").is_none());
+        assert!(json.get("raw_user_id").is_none());
+        assert_eq!(json["deviceId"], "DEVICE");
+        assert_eq!(json["state"], "keys_exchanged");
+        assert_eq!(json["createdAt"], 10);
+        assert_eq!(json["updatedAt"], 20);
+        assert_eq!(
+            json.pointer("/sas/decimals"),
+            Some(&json!([1234, 5678, 9012]))
+        );
+        assert_eq!(json.pointer("/sas/emoji/0/symbol"), Some(&json!("🐱")));
+        assert_eq!(
+            json.pointer("/sas/emoji/0/description"),
+            Some(&json!("Cat"))
+        );
+    }
+
+    #[test]
+    fn test_matrix_sas_info_wire_shape_omits_absent_optional_fields() {
+        let sas = MatrixSasInfo {
+            emoji: None,
+            decimals: Some([1, 2, 3]),
+        };
+        let json = serde_json::to_value(&sas).expect("serialize sas info");
+        assert_eq!(json, json!({ "decimals": [1, 2, 3] }));
+
+        let emoji = MatrixSasEmoji {
+            symbol: "7".to_string(),
+            description: "Seven".to_string(),
+        };
+        let json = serde_json::to_value(&emoji).expect("serialize sas emoji");
+        assert_eq!(json, json!({ "symbol": "7", "description": "Seven" }));
+    }
+
+    #[test]
+    fn test_verification_control_id_includes_user_and_protocol_flow() {
+        assert_eq!(
+            matrix_verification_control_id("@alice:example.com", "shared-protocol-flow"),
+            matrix_verification_control_id("@alice:example.com", "shared-protocol-flow"),
+            "daemon control ids must be deterministic for the same raw peer flow"
+        );
+        assert!(
+            matrix_verification_control_id("@alice:example.com", "shared-protocol-flow")
+                .starts_with("mvr_"),
+            "daemon control ids must live in the Matrix verification request namespace"
+        );
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (alice, inserted) = upsert_verification_record(
+            &state,
+            "shared-protocol-flow".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE1".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        let (bob, inserted) = upsert_verification_record(
+            &state,
+            "shared-protocol-flow".to_string(),
+            "@bob:example.com".to_string(),
+            Some("DEVICE2".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        assert_ne!(alice.flow_id, bob.flow_id);
+        assert_eq!(state.read().verifications.len(), 2);
+
+        let sanitized_collision = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let (raw_with_zwsp, inserted) = upsert_verification_record(
+            &sanitized_collision,
+            "flow-\u{200b}-id".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE1".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        let (raw_without_zwsp, inserted) = upsert_verification_record(
+            &sanitized_collision,
+            "flow--id".to_string(),
+            "@alice:example.com".to_string(),
+            Some("DEVICE1".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        assert_eq!(
+            raw_with_zwsp.protocol_flow_id,
+            raw_without_zwsp.protocol_flow_id
+        );
+        assert_ne!(
+            raw_with_zwsp.flow_id, raw_without_zwsp.flow_id,
+            "peer protocol ids that sanitize to the same display id must not collide in daemon control ids"
+        );
+        assert_eq!(sanitized_collision.read().verifications.len(), 2);
+    }
+
+    #[test]
+    fn test_verification_state_serde_round_trip_and_terminal() {
+        let encoded = serde_json::to_value(MatrixVerificationState::KeysExchanged).unwrap();
+        assert_eq!(encoded, json!("keys_exchanged"));
+        let decoded: MatrixVerificationState = serde_json::from_value(encoded).unwrap();
+        assert_eq!(decoded, MatrixVerificationState::KeysExchanged);
+        assert!(MatrixVerificationState::Done.is_terminal());
+        assert!(MatrixVerificationState::Cancelled.is_terminal());
+        assert!(MatrixVerificationState::Mismatched.is_terminal());
+        assert!(!MatrixVerificationState::Ready.is_terminal());
+    }
+
+    /// Verification-flow eviction must be terminal-first. Otherwise
+    /// a peer flooding fresh protocol_flow_ids fills the cap with
+    /// non-terminal records and evicts the operator's pending flow
+    /// at index 0 — denying the operator's verification UX.
+    #[test]
+    fn test_upsert_verification_record_evicts_terminal_first_at_cap() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        // Fill the cap with 1 OPERATOR-PENDING flow (non-terminal)
+        // followed by (CAP-1) PEER-INITIATED flows in mixed states,
+        // including some terminal so we can verify terminal-first.
+        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
+        // First record: a non-terminal flow at index 0 (the
+        // operator's). It must NOT be evicted on the next overflow.
+        let (_, inserted) = upsert_verification_record(
+            &state,
+            "operator-flow".to_string(),
+            "@operator:example.com".to_string(),
+            Some("OPERDEV".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        // Add cap-2 more non-terminals; record at index 1 will be a
+        // marker we can later spot when it's NOT evicted.
+        for i in 0..(cap - 2) {
+            upsert_verification_record(
+                &state,
+                format!("peer-flow-{i}"),
+                format!("@peer{i}:example.com"),
+                Some(format!("PEERDEV{i}")),
+                MatrixVerificationState::Requested,
+            )
+            .unwrap_applied();
+        }
+        // Add one TERMINAL record at the end (the SECOND-to-last,
+        // since we add one more terminal to ensure terminal eviction).
+        upsert_verification_record(
+            &state,
+            "old-terminal".to_string(),
+            "@cancelled:example.com".to_string(),
+            Some("DEVTERM".to_string()),
+            MatrixVerificationState::Cancelled,
+        )
+        .unwrap_applied();
+        // Cap is now reached. Insert one more — this triggers
+        // eviction. The OLDEST-TERMINAL is "old-terminal"; should be
+        // dropped. Operator's flow stays.
+        let (_, inserted) = upsert_verification_record(
+            &state,
+            "trigger-evict".to_string(),
+            "@new:example.com".to_string(),
+            Some("DEVNEW".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        assert!(inserted);
+        let guard = state.read();
+        let flow_ids: Vec<&str> = guard
+            .verifications
+            .iter()
+            .map(|f| f.protocol_flow_id.as_str())
+            .collect();
+        assert!(
+            flow_ids.contains(&"operator-flow"),
+            "operator's pending flow MUST NOT be evicted under peer flood"
+        );
+        assert!(
+            !flow_ids.contains(&"old-terminal"),
+            "oldest terminal record MUST be evicted before any non-terminal"
+        );
+    }
+
+    /// Multi-call cumulative cap-eviction with a terminal-rich
+    /// stream. As long as terminal records exist in the buffer at
+    /// eviction time, the operator's non-terminal flow at index 0
+    /// must NEVER be evicted. The all-non-terminal case is the
+    /// documented backstop scenario where TTL pruning protects
+    /// the flow; this test explicitly avoids that scenario by
+    /// keeping the buffer terminal-rich (every insert is terminal
+    /// after the operator's flow goes in).
+    #[test]
+    fn test_upsert_verification_record_multi_call_cumulative_eviction_preserves_operator_flow() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
+
+        upsert_verification_record(
+            &state,
+            "operator-flow".to_string(),
+            "@operator:example.com".to_string(),
+            Some("OPERDEV".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+
+        // Drive 3*cap upserts of TERMINAL records (peer-initiated
+        // verifications that immediately resolve to Cancelled).
+        // Each insert above cap triggers eviction; with only
+        // terminals available, the eviction policy drops the
+        // oldest terminal and never touches the operator's
+        // non-terminal at index 0.
+        for i in 0..(3 * cap) {
+            upsert_verification_record(
+                &state,
+                format!("peer-flow-{i}"),
+                format!("@peer{i}:example.com"),
+                Some(format!("DEV{i}")),
+                MatrixVerificationState::Cancelled,
+            )
+            .unwrap_applied();
+        }
+
+        let guard = state.read();
+        assert!(
+            guard.verifications.len() <= cap,
+            "verification cap must hold under sustained upserts: got {} > {cap}",
+            guard.verifications.len(),
+        );
+        let has_operator = guard
+            .verifications
+            .iter()
+            .any(|f| f.protocol_flow_id == "operator-flow");
+        assert!(
+            has_operator,
+            "operator's pending flow at index 0 must survive cumulative terminal-rich \
+             cap eviction — terminal-first eviction policy must protect index 0"
+        );
+    }
+
+    /// All-terminal cap-eviction: when every record in the buffer
+    /// is terminal, eviction drops the oldest terminal (which is
+    /// the oldest record overall — same outcome). Pin that the
+    /// eviction does NOT panic and that the buffer stays bounded.
+    #[test]
+    fn test_upsert_verification_record_all_terminal_at_cap_drops_oldest() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
+        for i in 0..(cap + 5) {
+            upsert_verification_record(
+                &state,
+                format!("flow-{i}"),
+                format!("@user{i}:example.com"),
+                Some(format!("DEV{i}")),
+                MatrixVerificationState::Cancelled,
+            )
+            .unwrap_applied();
+        }
+        let guard = state.read();
+        assert_eq!(guard.verifications.len(), cap);
+        // Oldest 5 records (flow-0 through flow-4) should be
+        // evicted; flow-5 onward should remain.
+        for i in 0..5 {
+            let id = format!("flow-{i}");
+            assert!(
+                !guard.verifications.iter().any(|f| f.protocol_flow_id == id),
+                "flow-{i} (oldest) should have been evicted"
+            );
+        }
+        for i in 5..(cap + 5) {
+            let id = format!("flow-{i}");
+            assert!(
+                guard.verifications.iter().any(|f| f.protocol_flow_id == id),
+                "flow-{i} (recent) should still be in the buffer"
+            );
+        }
+    }
+
+    #[test]
+    fn test_upsert_verification_record_refuses_when_all_records_are_active() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
+        for i in 0..cap {
+            upsert_verification_record(
+                &state,
+                format!("active-flow-{i}"),
+                format!("@user{i}:example.com"),
+                Some(format!("DEV{i}")),
+                MatrixVerificationState::KeysExchanged,
+            )
+            .unwrap_applied();
+        }
+
+        let result = upsert_verification_record(
+            &state,
+            "new-peer-flow".to_string(),
+            "@new:example.com".to_string(),
+            Some("NEWDEV".to_string()),
+            MatrixVerificationState::Requested,
+        );
+
+        assert!(matches!(result, VerificationRecordUpsert::RejectedAtCap));
+        let guard = state.read();
+        assert_eq!(guard.verifications.len(), cap);
+        assert!(
+            guard
+                .verifications
+                .iter()
+                .any(|f| f.protocol_flow_id == "active-flow-0"),
+            "active ceremonies must not be evicted to admit a fresh peer request"
+        );
+        assert!(
+            !guard
+                .verifications
+                .iter()
+                .any(|f| f.protocol_flow_id == "new-peer-flow"),
+            "fresh peer request must not be recorded when every capped slot is active"
+        );
+    }
+
+    #[test]
+    fn test_upsert_verification_record_same_peer_eviction_uses_raw_user_id() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let raw_plain = "@alice:example.com".to_string();
+        let raw_with_zwsp = "@ali\u{200b}ce:example.com".to_string();
+
+        upsert_verification_record(
+            &state,
+            "victim-flow".to_string(),
+            raw_plain.clone(),
+            Some("VICTIM".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+
+        for i in 1..MATRIX_VERIFICATION_RECORDS_MAX {
+            upsert_verification_record(
+                &state,
+                format!("active-flow-{i}"),
+                format!("@peer{i}:example.com"),
+                Some(format!("ACTIVE{i}")),
+                MatrixVerificationState::KeysExchanged,
+            )
+            .unwrap_applied();
+        }
+
+        let result = upsert_verification_record(
+            &state,
+            "incoming-flow".to_string(),
+            raw_with_zwsp.clone(),
+            Some("INCOMING".to_string()),
+            MatrixVerificationState::Requested,
+        );
+
+        assert!(matches!(result, VerificationRecordUpsert::RejectedAtCap));
+        let guard = state.read();
+        assert!(
+            guard.verifications.iter().any(|flow| {
+                flow.raw_user_id == raw_plain && flow.protocol_flow_id == "victim-flow"
+            }),
+            "sanitized-colliding raw user IDs must not evict each other's requested flows"
+        );
+        assert!(
+            !guard
+                .verifications
+                .iter()
+                .any(|flow| flow.raw_user_id == raw_with_zwsp),
+            "new flow must be rejected rather than evicting a different raw user's flow"
+        );
+    }
+
+    /// Exactly-at-cap: no eviction happens on the boundary insert.
+    /// The cap is `MATRIX_VERIFICATION_RECORDS_MAX` items; the
+    /// `cap+1`-th insert triggers the first eviction. Pin that
+    /// exactly-cap is steady-state, no flapping.
+    #[test]
+    fn test_upsert_verification_record_at_cap_does_not_evict() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        let cap = MATRIX_VERIFICATION_RECORDS_MAX;
+        for i in 0..cap {
+            upsert_verification_record(
+                &state,
+                format!("flow-{i}"),
+                format!("@user{i}:example.com"),
+                Some(format!("DEV{i}")),
+                MatrixVerificationState::Requested,
+            )
+            .unwrap_applied();
+        }
+        let guard = state.read();
+        assert_eq!(
+            guard.verifications.len(),
+            cap,
+            "exactly-cap insert count must produce exactly-cap buffer length"
+        );
+        // All flows must be present (no evictions yet).
+        for i in 0..cap {
+            let id = format!("flow-{i}");
+            assert!(
+                guard.verifications.iter().any(|f| f.protocol_flow_id == id),
+                "flow-{i} must still be in the buffer at exactly-cap"
+            );
+        }
+    }
+
+    /// Verify-action against a flow already in `Cancelled`/`Done`/
+    /// `Mismatched` must surface `MatrixError::VerificationCancelled`,
+    /// NOT a generic Verification error or a state-transition success.
+    /// The CLI routes 410 Gone on this variant, signaling the operator
+    /// to start a new flow rather than retry. This test pins the
+    /// classifier's terminal-state guard at the helper level by
+    /// stamping a verification record into runtime state and reading
+    /// back the post-cancel snapshot.
+    #[test]
+    fn test_upsert_verification_record_preserves_terminal_state_after_cancel() {
+        let state = Arc::new(RwLock::new(MatrixRuntimeState::default()));
+        // Initial flow: Requested.
+        upsert_verification_record(
+            &state,
+            "flow-1".to_string(),
+            "@peer:example.com".to_string(),
+            Some("DEVPEER".to_string()),
+            MatrixVerificationState::Requested,
+        )
+        .unwrap_applied();
+        // Update to Cancelled (terminal). State machine must store
+        // the terminal state so a later Confirm can see it.
+        let (rec, _) = upsert_verification_record(
+            &state,
+            "flow-1".to_string(),
+            "@peer:example.com".to_string(),
+            Some("DEVPEER".to_string()),
+            MatrixVerificationState::Cancelled,
+        )
+        .unwrap_applied();
+        assert_eq!(rec.state, MatrixVerificationState::Cancelled);
+
+        // The exact `apply_verification_action` plumbing requires
+        // an SDK `VerificationRequest` handle; the testable surface
+        // is the state lookup that `apply_verification_action`
+        // would consult. Pin that the cancelled record is
+        // discoverable by `protocol_flow_id` so the action path
+        // can return `VerificationCancelled`.
+        let guard = state.read();
+        let found = guard
+            .verifications
+            .iter()
+            .find(|f| f.protocol_flow_id == "flow-1")
+            .expect("flow-1 must exist");
+        assert!(found.state.is_terminal());
+        assert_eq!(found.state, MatrixVerificationState::Cancelled);
+    }
+
+    #[test]
+    fn test_verification_terminal_guard_rejects_accept_and_confirm_but_allows_cancel() {
+        for terminal_state in [
+            MatrixVerificationState::Cancelled,
+            MatrixVerificationState::Done,
+            MatrixVerificationState::Mismatched,
+        ] {
+            for action in [
+                MatrixVerificationAction::Accept,
+                MatrixVerificationAction::Confirm { matches: true },
+            ] {
+                let err =
+                    guard_verification_action_terminal_state("flow-1", &action, &terminal_state)
+                        .expect_err(
+                        "accept/confirm against terminal flow must be rejected before SDK lookup",
+                    );
+                assert!(matches!(
+                    err,
+                    MatrixError::VerificationCancelled {
+                        flow_id,
+                        state,
+                    } if flow_id == "flow-1" && state == terminal_state
+                ));
+            }
+
+            guard_verification_action_terminal_state(
+                "flow-1",
+                &MatrixVerificationAction::Cancel,
+                &terminal_state,
+            )
+            .expect("cancel against terminal flow is idempotent");
+        }
+    }
+
+    /// `to-device` verification events from neither the
+    /// configured user nor the auto-join allowlist must be dropped
+    /// without entering the verification record store. Otherwise a
+    /// hostile peer can spam 256 fresh transaction_ids and evict
+    /// the operator's legitimate flow at index 0 (the cap-eviction
+    /// fallback path).
+    ///
+    /// The handler is async and consumes an SDK event type that's
+    /// awkward to construct in unit tests, so this test exercises
+    /// the GATE policy directly via the helper
+    /// `MatrixAutoJoinConfig::allows_user` and the self-equality
+    /// check that the handler uses. A regression that loosens the
+    /// gate (e.g. removes the `allows_user` arm) would not trip
+    /// this test, but a regression that breaks the helpers
+    /// themselves would, and a static-analysis pin against the
+    /// handler body catches the gate-removal class of refactor.
+    #[test]
+    fn test_handle_to_device_verification_gate_helpers_reject_unallowed_peer() {
+        let config = matrix_test_config(false);
+        // Non-allowlisted peer, not the configured user.
+        assert!(
+            !config.auto_join.allows_user("@hostile-peer:evil.com"),
+            "default test config must not allowlist arbitrary peers"
+        );
+        // Configured user is `@cara:example.com` per matrix_test_config.
+        assert_eq!(config.user_id, "@cara:example.com");
+
+        // Pin the source body — the handler's gate must combine
+        // self-equality OR allowlist; a refactor that drops either
+        // arm breaks the contract.
+        let body = matrix_rs_fn_body("async fn handle_to_device_event");
+        let body = body.as_str();
+        assert!(
+            body.contains("matrix_user_ids_equal(sender, &config.user_id)")
+                || body.contains("matrix_user_ids_equal(\n"),
+            "handle_to_device_event must check self-verification via matrix_user_ids_equal"
+        );
+        assert!(
+            body.contains("auto_join.allows_user"),
+            "handle_to_device_event must consult the allowlist for non-self peers"
+        );
+        assert!(
+            body.contains("to-device verification event dropped"),
+            "handle_to_device_event must drop unallowed peers (warn-log marker present)"
+        );
+        let classifier = matrix_rs_fn_body("fn matrix_to_device_verification_sender_and_kind");
+        let classifier = classifier.as_str();
+        for event_type in [
+            "m.key.verification.start",
+            "m.key.verification.ready",
+            "m.key.verification.key",
+            "m.key.verification.mac",
+        ] {
+            assert!(
+                classifier.contains(event_type),
+                "to-device verification classifier must include {event_type}"
+            );
+        }
+    }
 }
