+++ src/server/ws/tests.rs
@@ -1257,6 +1257,7 @@ fn test_matrix_verification_request_rate_class_from_info_routes_through_is_termi
         protocol_flow_id: "proto-flow".to_string(),
         raw_protocol_flow_id: "proto-flow".to_string(),
         user_id: "@alice:example.com".to_string(),
+        raw_user_id: "@alice:example.com".to_string(),
         device_id: Some("DEVICE".to_string()),
         state,
         sas: None,
@@ -1318,6 +1319,7 @@ fn test_matrix_verification_request_rate_key_from_info_matches_value_path() {
             protocol_flow_id: format!("{flow_id}-protocol"),
             raw_protocol_flow_id: format!("{flow_id}-protocol"),
             user_id: user_id.to_string(),
+            raw_user_id: user_id.to_string(),
             device_id: device_id.map(String::from),
             state: MatrixVerificationState::Requested,
             sas: None,
@@ -4982,6 +4984,7 @@ fn test_new_verification_flow_from_upsert_gates_on_inserted() {
         protocol_flow_id: "txn".to_string(),
         raw_protocol_flow_id: "txn".to_string(),
         user_id: "@a:x".to_string(),
+        raw_user_id: "@a:x".to_string(),
         device_id: None,
         state: MatrixVerificationState::Requested,
         sas: None,
@@ -5052,6 +5055,7 @@ fn test_broadcast_matrix_verification_request_delivers_only_to_admin_scope() {
         protocol_flow_id: "txn-abc".to_string(),
         raw_protocol_flow_id: "txn-abc".to_string(),
         user_id: "@alice:example.com".to_string(),
+        raw_user_id: "@alice:example.com".to_string(),
         device_id: Some("DEVICE".to_string()),
         state: MatrixVerificationState::Requested,
         sas: None,
